refactor(static): wip 1
