--- static/assets/js/icons.dist.js
@@ -1,25 +1,0 @@
-Iconify.addCollection({"prefix":"fxemoji","icons":{"key":{"body":"<path fill=\"#FFB636\" d=\"M456.943 48.057c-64.075-64.075-167.962-64.075-232.038 0c-50.857 50.857-61.341 126.792-31.466 187.997l-45.691 45.691a3.594 3.594 0 0 0-1.052 2.59l.403 28.553a3.59 3.59 0 0 1-1.052 2.59l-23.879 23.879a3.591 3.591 0 0 1-3.249.981l-22.888-4.61a3.594 3.594 0 0 0-3.249.981L67.443 362.05a3.59 3.59 0 0 0-1.047 2.357l-1.458 28.668a3.592 3.592 0 0 1-3.881 3.397l-24.246-1.996a3.59 3.59 0 0 0-2.834 1.04l-22.05 22.05a3.59 3.59 0 0 0 0 5.079l10.938 10.937a3.59 3.59 0 0 1 0 5.079L14.526 447a3.595 3.595 0 0 0-1.051 2.456l-1.146 49.15a3.591 3.591 0 0 0 1.051 2.623l7.793 7.793c.72.72 1.711 1.1 2.727 1.047l47.721-2.49a3.586 3.586 0 0 0 2.352-1.047l194.973-194.973c61.205 29.875 137.14 19.391 187.997-31.466c64.076-64.074 64.076-167.961 0-232.036zm-23.812 76.438c-14.532 14.532-38.094 14.532-52.626 0s-14.532-38.094 0-52.626c14.532-14.532 38.094-14.532 52.626 0c14.532 14.532 14.532 38.094 0 52.626z\"/>"},"rocket":{"body":"<path fill=\"#FFB636\" d=\"m204.854 309.75l.668 1.748c.208.564.522 1.32.791 2.115l.971 2.795c.38 1.069.689 2.149 1.094 3.4c.386 1.234.843 2.615 1.201 3.973l1.25 4.47c.412 1.558.758 3.122 1.175 4.826a162.186 162.186 0 0 1 3.798 22.366c.426 4.01.607 8.007.696 12.117c.006 4.028-.091 8.159-.468 12.194a101.15 101.15 0 0 1-1.825 11.94a84.231 84.231 0 0 1-3.351 11.379c-2.757 7.354-6.624 14.15-11.647 20.157c-2.485 3.031-5.389 5.734-8.266 8.533c-2.918 2.758-5.904 5.517-9.021 8.191a238.747 238.747 0 0 1-19.729 15.385c-6.908 4.858-14.174 9.358-21.627 13.487a259.059 259.059 0 0 1-11.347 5.865a260.232 260.232 0 0 1-11.582 5.217c-7.784 3.247-15.627 6.066-23.403 8.401a238.5 238.5 0 0 1-22.815 5.681c-14.714 2.934-28.219 4.137-38.257 4.394l-3.589.062l-3.223-.003l-5.218-.139l-3.345-.196l-1.182-.102l-.102-1.182l-.196-3.345l-.139-5.218l-.003-3.223l.062-3.589c.257-10.038 1.459-23.543 4.394-38.257a238.936 238.936 0 0 1 5.68-22.815c2.335-7.776 5.154-15.62 8.401-23.403a260.75 260.75 0 0 1 5.217-11.582c1.86-3.816 3.81-7.611 5.865-11.347c4.129-7.453 8.629-14.719 13.487-21.627a238.532 238.532 0 0 1 15.385-19.729c2.674-3.117 5.432-6.104 8.191-9.021c2.799-2.877 5.502-5.781 8.533-8.266c6.008-5.023 12.803-8.89 20.157-11.647a84.376 84.376 0 0 1 11.379-3.351a101.15 101.15 0 0 1 11.94-1.825c4.035-.377 8.167-.474 12.194-.468c4.11.088 8.106.27 12.117.696a162.035 162.035 0 0 1 22.366 3.798c1.703.416 3.267.763 4.826 1.175l4.47 1.25c1.358.359 2.74.815 3.973 1.201c1.252.404 2.331.714 3.4 1.094l2.795.971c.795.269 1.551.583 2.115.791c1.149.435 1.744.663 1.744.663z\"/><path fill=\"#FF473E\" d=\"M76.86 181.752L4.346 254.266c-5.012 5.012-1.462 13.582 5.626 13.582h60.945c2.11 0 4.134.838 5.626 2.33l39.572 39.572l64.271-130.328h-97.9a7.952 7.952 0 0 0-5.626 2.33zm139.301 210.84l39.572 39.572a7.957 7.957 0 0 1 2.33 5.626v60.945c0 7.088 8.57 10.638 13.582 5.626l72.514-72.514a7.957 7.957 0 0 0 2.33-5.626v-97.9l-130.328 64.271z\"/><path fill=\"#ADBBBC\" d=\"m505.647 17.875l.077.637c.047.419.139 1.065.205 1.862c.131 1.593.428 4.061.569 7.033c.452 6.114.756 14.721.543 25.076c-.207 10.361-.937 22.464-2.519 35.602c-.354 3.326-.859 6.619-1.347 10.041c-.453 3.457-1.098 6.834-1.687 10.373c-1.21 7.047-2.72 14.218-4.463 21.533a361.335 361.335 0 0 1-13.805 44.697c-5.738 15.021-12.624 30.025-20.695 44.6c-8.083 14.563-17.278 28.768-27.555 42.27c-10.263 13.515-21.594 26.34-33.751 38.338c-12.157 11.999-25.124 23.187-38.748 33.341c-3.407 2.537-6.834 5.032-10.338 7.419c-3.466 2.426-7.007 4.746-10.536 7.044c-7.092 4.562-14.334 8.831-21.651 12.837a361.966 361.966 0 0 1-44.669 20.626a372.522 372.522 0 0 1-44.644 13.858c-7.299 1.759-14.432 3.307-21.467 4.528l-5.179.933c-1.716.294-3.457.536-5.147.801l-5.015.765l-4.974.633c-13.067 1.653-25.09 2.463-35.375 2.746c-10.274.294-18.832.04-24.876-.343c-2.942-.111-5.394-.393-6.968-.504c-.789-.058-1.43-.145-1.844-.187l-.631-.07c-4.782-.533-8.409-4.329-8.919-8.919l-.07-.63c-.042-.414-.129-1.055-.187-1.844c-.111-1.574-.393-4.025-.504-6.968c-.382-6.044-.636-14.601-.343-24.876c.283-10.285 1.093-22.308 2.746-35.375l.633-4.974l.765-5.015c.265-1.69.508-3.43.801-5.147l.933-5.179c1.221-7.035 2.769-14.168 4.528-21.467a372.104 372.104 0 0 1 13.858-44.644a362.26 362.26 0 0 1 20.626-44.669c4.007-7.316 8.276-14.559 12.837-21.651c2.298-3.529 4.618-7.07 7.044-10.536c2.387-3.505 4.882-6.931 7.419-10.338c10.154-13.624 21.342-26.592 33.341-38.748c11.998-12.157 24.823-23.488 38.338-33.751c13.502-10.276 27.707-19.472 42.27-27.555c14.575-8.071 29.58-14.957 44.6-20.695a361.335 361.335 0 0 1 44.697-13.805c7.315-1.743 14.486-3.253 21.533-4.463c3.539-.589 6.916-1.234 10.373-1.687c3.422-.488 6.714-.994 10.041-1.347c13.138-1.582 25.241-2.312 35.602-2.519c10.355-.213 18.962.09 25.076.543c2.972.141 5.44.439 7.033.569c.797.066 1.443.157 1.862.205l.637.077c4.769.575 8.37 4.354 8.92 8.919z\"/><path fill=\"#FF473E\" d=\"m243.106 271.497l.583 1.577c.325.981.763 2.392 1.065 3.964c.716 3.255.953 7.217-.341 10.663c-.638 1.732-1.818 3.176-2.929 4.922l-3.693 5.555l-8.865 13.312a1631.323 1631.323 0 0 1-10.652 15.587a1262.693 1262.693 0 0 1-12.115 17.172l-6.51 8.959l-6.764 9.085c-4.589 6.076-9.285 12.215-14.06 18.274a1161.386 1161.386 0 0 1-14.364 17.97c-4.781 5.885-9.469 11.692-14.083 17.235l-6.769 8.192c-2.242 2.646-4.281 5.389-6.649 7.677c-2.353 2.302-5.068 4.116-8.188 5.376c-3.093 1.287-6.479 2.133-9.753 2.923a195.747 195.747 0 0 1-18.695 3.482c-2.904.397-5.681.709-8.304.944c-2.609.248-5.133.347-7.345.506c-2.229.141-4.357.129-6.135.192c-1.805.036-3.43-.023-4.702-.025c-1.299-.029-2.337-.094-3.016-.117l-1.066-.072l-.072-1.066c-.023-.679-.088-1.718-.117-3.016c-.002-1.272-.061-2.896-.025-4.702c.062-1.779.051-3.906.192-6.135c.158-2.212.258-4.736.506-7.345c.234-2.622.546-5.4.944-8.304a195.747 195.747 0 0 1 3.482-18.695c.79-3.273 1.636-6.659 2.923-9.753c1.26-3.121 3.074-5.835 5.376-8.188c2.288-2.367 5.031-4.406 7.678-6.648l8.192-6.769c5.543-4.614 11.35-9.303 17.235-14.083c5.875-4.79 11.893-9.607 17.97-14.364c6.059-4.776 12.198-9.471 18.274-14.06l9.085-6.764l8.959-6.51a1271.61 1271.61 0 0 1 17.172-12.115c5.5-3.811 10.739-7.375 15.587-10.652l13.312-8.865l5.555-3.693c1.745-1.111 3.189-2.291 4.922-2.929c3.446-1.294 7.408-1.057 10.663-.341c1.572.302 2.983.74 3.964 1.065l1.573.579z\"/><path fill=\"#2B3B47\" d=\"M362.113 206.66c0 29.917-24.253 54.17-54.17 54.17s-54.17-24.253-54.17-54.17s24.253-54.17 54.17-54.17s54.17 24.253 54.17 54.17zm27.229-115.881c-19.044 0-34.482 15.438-34.482 34.482s15.438 34.482 34.482 34.482s34.482-15.438 34.482-34.482s-15.438-34.482-34.482-34.482z\"/>"},"satelliteantenna":{"body":"<path fill=\"#ADB8BC\" d=\"M276.666 473.233c.007.178.02.356.02.534c0 20.563-49.116 37.233-109.704 37.233c-60.588 0-109.704-16.67-109.704-37.233c0-.179.013-.356.02-.534c-.759-3.345-1.159-7.642-1.159-13.089c0-61.217 49.626-110.843 110.843-110.843s110.843 49.626 110.843 110.843c0 5.446-.4 9.744-1.159 13.089z\"/><path fill=\"#C5CFD1\" d=\"M255.296 226.867c69.528 69.528 166.832 84.952 104.321 147.462S190.068 431.15 120.54 361.622s-75.217-176.567-12.706-239.077s77.934 34.793 147.462 104.322z\"/><path fill=\"#6C7B7F\" d=\"M309.824 172.339c69.528 69.528 96.116 155.668 59.385 192.399s-122.871 10.143-192.399-59.385s-96.116-155.668-59.385-192.399s122.871-10.143 192.399 59.385z\"/><path fill=\"#C5CFD1\" d=\"M368.513 165.432c-10.408 10.408-25.522 13.232-38.492 8.488L216.213 287.728a15.353 15.353 0 0 1-10.889 4.511a15.348 15.348 0 0 1-10.889-4.511c-6.014-6.014-6.014-15.764 0-21.777l113.808-113.808c-4.745-12.97-1.921-28.084 8.487-38.492c14.299-14.299 37.483-14.299 51.782 0s14.3 37.482.001 51.781z\"/><path fill=\"#FFB636\" d=\"M414.738 161.526c-.641 0-1.289-.051-1.941-.158c-6.542-1.064-10.981-7.23-9.918-13.771a60.12 60.12 0 0 0-1.689-26.607c-8.347-28.107-37.607-46.425-66.625-41.707c-6.538 1.062-12.706-3.375-13.771-9.917c-1.063-6.542 3.376-12.707 9.918-13.771c19.83-3.226 40.305.974 57.656 11.825c17.383 10.87 30.107 27.469 35.83 46.738c3.566 12.009 4.386 24.905 2.371 37.293c-.959 5.889-6.052 10.075-11.831 10.075zm62.063-12.28c3.027-18.598 1.797-37.958-3.558-55.989c-8.583-28.901-27.684-53.808-53.785-70.13c-26.06-16.296-56.795-22.605-86.54-17.766c-6.542 1.064-10.981 7.23-9.918 13.771c1.065 6.542 7.233 10.979 13.771 9.917c23.995-3.905 48.842 1.219 69.962 14.426c21.137 13.218 36.587 33.324 43.502 56.613c4.332 14.586 5.327 30.25 2.878 45.301c-1.065 6.542 3.375 12.708 9.916 13.772c.653.106 1.302.158 1.942.158c5.779 0 10.872-4.185 11.83-10.073z\"/><path fill=\"#6C7B7F\" d=\"M276.666 473.232c.007.178.02.356.02.534c0 20.563-49.116 37.233-109.704 37.233c-60.588 0-109.704-16.67-109.704-37.233c0-.179.013-.356.02-.534c-.759-3.345-1.159-7.642-1.159-13.089c0-3.62.179-7.197.518-10.728c.171 1.347.386 2.575.641 3.697c-.007.178-.02.356-.02.534c0 3.486 1.416 6.86 4.055 10.06c4.094 4.962 11.142 9.505 20.459 13.398c20.114 8.404 50.802 13.776 85.19 13.776s65.076-5.372 85.19-13.776c9.317-3.893 16.365-8.435 20.459-13.398c2.64-3.2 4.055-6.574 4.055-10.06c0-.179-.013-.356-.02-.534c.255-1.122.469-2.351.641-3.697c.339 3.531.518 7.108.518 10.728c0 5.447-.4 9.745-1.159 13.089z\"/>"},"lockandkey":{"body":"<path fill=\"#B1B4B5\" d=\"M278.297 352.333c-11.783 0-21.335-9.553-21.335-21.336V206.76c0-41.873-34.066-75.939-75.939-75.939c-41.874 0-75.94 34.066-75.94 75.939v72.259c0 11.783-9.552 21.336-21.335 21.336s-21.335-9.553-21.335-21.336V206.76c0-65.402 53.208-118.61 118.611-118.61c65.402 0 118.61 53.208 118.61 118.61v124.237c-.002 11.783-9.554 21.336-21.337 21.336z\"/><path fill=\"#FFB636\" d=\"M304.326 481.531H57.717c-27.835 0-50.4-22.565-50.4-50.4V294.048c0-27.835 22.565-50.4 50.4-50.4h246.609c27.835 0 50.4 22.565 50.4 50.4v137.084c0 27.834-22.565 50.399-50.4 50.399z\"/><path fill=\"#FFD469\" d=\"M46.296 459.982H42.31c-6.432 0-11.647-5.214-11.647-11.647V304.487c0-6.432 5.214-11.646 11.647-11.646h3.986c6.432 0 11.647 5.214 11.647 11.646v143.848c-.001 6.433-5.215 11.647-11.647 11.647z\"/><path fill=\"#597B91\" d=\"M511 154.03c0-56.235-45.588-101.823-101.823-101.823c-56.235 0-101.823 45.588-101.823 101.823c0 44.893 29.053 82.997 69.382 96.543l-.133 9.098l-.155 10.673l-.079 5.52l-.014 5.626l-.06 23.333c-.013 7.946.13 16.018.188 24.09l.117 12.092l.06 6.017l.132 5.981l.516 23.333l.247 11.146c.036 1.821.073 3.622.172 5.402c.139 1.779.294 3.537.593 5.272c.285 1.734.624 3.445 1.092 5.13l.677 2.508l.788 2.468c1.027 3.263 2.08 6.417 2.996 9.444c.922 3.027 1.681 5.928 2.786 8.687c1.141 2.759 2.354 5.376 3.528 7.836c2.343 4.919 4.318 9.207 6.119 12.739l2.349 4.714l1.811 3.484c.486.946.873 1.671 1.141 2.16l.4.741c.141.259.302.532.467.783c3.2 4.873 9.745 6.229 14.618 3.029l5.804-3.812l.285-.187c.197-.124.47-.309.801-.554l1.263-.908c.457-.361 1.024-.778 1.586-1.251c2.31-1.892 5.352-4.667 8.232-8.198c2.909-3.531 5.746-7.82 7.782-12.739c.441-1.23 1.231-2.499.885-3.805c-.259-1.307-1.563-2.651-3.152-4.03c-1.591-1.379-3.193-2.794-3.781-4.243c-.334-.724-.341-1.457-.367-2.198c-.034-.741-.069-1.49-.042-2.246c.023-2.628.354-5.357.954-8.165c-6.199-.944-10.953-6.281-10.953-12.744c0-6.024 4.133-11.07 9.714-12.493c.947.241 1.933.383 2.955.383c6.601 0 11.952-5.351 11.952-11.951v-36.173c0-5.828-4.174-10.676-9.695-11.732l.001-.224h-2.019c-7.129 0-10.996-5.779-10.996-12.908c0-6.145 2.88-11.278 8.239-12.58h1.321c.945 0 1.859-.121 2.741-.328h.748l-.001-.192c4.899-1.493 8.464-6.045 8.464-11.432c0-5.416-3.606-9.985-8.547-11.452c-.041-5.522-.126-10.887-.168-16.043l-.095-8.897C481.649 237.41 511 199.151 511 154.03zm-45.654 0H353.008c-8.109 0-14.178-7.508-12.406-15.42c7.029-31.393 35.062-54.852 68.575-54.852s61.547 23.459 68.575 54.852c1.772 7.913-4.297 15.42-12.406 15.42z\"/><path fill=\"#7E9AA8\" d=\"M396.27 379.705a7.927 7.927 0 0 1-7.926-7.926V266.204a7.927 7.927 0 0 1 15.852 0v105.575a7.928 7.928 0 0 1-7.926 7.926z\"/>"},"clipboard":{"body":"<path fill=\"#D19B61\" d=\"M402.634 505.687H109.366c-19.882 0-36-16.118-36-36V93.822c0-19.882 16.118-36 36-36h293.267c19.882 0 36 16.118 36 36v375.864c.001 19.883-16.117 36.001-35.999 36.001z\"/><path fill=\"#F9F9F7\" d=\"M386.152 463.673H125.848c-7.953 0-14.4-6.447-14.4-14.4V102.306c0-7.953 6.447-14.4 14.4-14.4h260.305c7.953 0 14.4 6.447 14.4 14.4v346.967c-.001 7.953-6.448 14.4-14.401 14.4z\"/><path fill=\"#A9B8C2\" d=\"M340.027 200.306H153.508a7.772 7.772 0 1 1 0-15.544h186.519a7.772 7.772 0 1 1 0 15.544zm-76.255 36.065a7.772 7.772 0 0 0-7.772-7.772H153.508a7.772 7.772 0 1 0 0 15.544H256a7.772 7.772 0 0 0 7.772-7.772zm0-88.016a7.772 7.772 0 0 0-7.772-7.772H153.508a7.772 7.772 0 1 0 0 15.544H256a7.772 7.772 0 0 0 7.772-7.772zM347.8 280.208a7.772 7.772 0 0 0-7.772-7.772h-186.52a7.772 7.772 0 0 0 0 15.544h186.519a7.773 7.773 0 0 0 7.773-7.772zm-55.956 43.838a7.772 7.772 0 0 0-7.772-7.772H153.508a7.772 7.772 0 0 0 0 15.544h130.563a7.773 7.773 0 0 0 7.773-7.772zm55.956 43.837a7.772 7.772 0 0 0-7.772-7.772h-186.52a7.772 7.772 0 0 0 0 15.544h186.519a7.773 7.773 0 0 0 7.773-7.772zM226.562 411.72a7.772 7.772 0 0 0-7.772-7.772h-65.282a7.772 7.772 0 0 0 0 15.544h65.282a7.772 7.772 0 0 0 7.772-7.772z\"/><path fill=\"#B9C5C6\" d=\"M307.042 35.654h-15.388C291.654 15.963 275.691 0 256 0s-35.654 15.963-35.654 35.654h-15.388c-19.882 0-36 16.118-36 36v36.98a7.2 7.2 0 0 0 7.2 7.2h159.685a7.2 7.2 0 0 0 7.2-7.2v-36.98c-.001-19.882-16.118-36-36.001-36z\"/><circle cx=\"256\" cy=\"35.654\" r=\"23.41\" fill=\"#2B3B47\"/>"},"running":{"body":"<path fill=\"#FFD3B6\" d=\"M302.3 101.7c-6.3-1.2-33.4-11.9-33.4-23.3c0-11.1-9.8-11.7-12.3-11.7c-3.4 0-5.1-2.4-5.1-5.8c0-43.3-27.4-31.2-28.9-31.2c-2.8 0 0-6-2.4-6c-.8 0-1.5-.1-2.2-.1c-6.8 7.2-10.7 17.2-10.7 29.7l.1 6.9v.1c-.1 1.3-4.5 14.8-7.3 20.7c-3.2 6.6 1 8 7.8 8.2l.2 11.2c0 2.1 5.9 1.8 7.9.7c1.5-.8 3.4-.2 4.2 1.3c.8 1.5.2 3.4-1.3 4.2c-1.7.9-5.1 1.5-7.5 1.9c-1.3.2-2.2 1.6-1.9 2.9c3.9 13.1 15.4 18.5 30.4 20.4v15.8c0 5.3 8.6 9.6 19.2 9.6s19.2-4.3 19.2-9.6c0-.6-.1-1.2-.3-1.7v-17.4c11.9-5.7 19.5-15.9 24.3-26.8z\"/><path fill=\"#2B3B47\" d=\"M226.9 65c-.2-2.3-2.1-4-4.4-4c-2.3 0-4.2 1.8-4.4 4h-.1v6.2c0 2.4 2 4.4 4.4 4.4c2.4 0 4.4-2 4.4-4.4l.1-6.2z\"/><path fill=\"#FFF\" d=\"M221.6 65.2c0-.6-.5-1-1.1-1c-.6 0-1 .4-1.1 1v1.5c0 .6.5 1.1 1.1 1.1c.6 0 1.1-.5 1.1-1.1v-1.5z\"/><path fill=\"#E5AB83\" d=\"M215.3 51.5s.2 0 .7-.1c.4 0 1-.1 1.7-.1s1.5-.1 2.4-.1c.9 0 1.9.1 2.8.2c1 .1 1.9.3 2.9.5c.5.1.9.3 1.3.4c.4.1.8.3 1.2.5c.4.1.7.4 1 .6c.3.2.5.4.7.5c.4.3.6.5.6.5c.6.6.7 1.5.1 2.2c-.5.6-1.4.7-2 .2l-.3-.2s-.1-.1-.4-.3c-.1-.1-.2-.2-.4-.3c-.2-.1-.4-.2-.7-.3c-.3-.1-.5-.2-.9-.3c-.3-.1-.7-.2-1-.2c-.7-.1-1.5-.2-2.4-.2c-.8 0-1.7 0-2.5.1s-1.5.1-2.2.2c-.6.1-1.2.2-1.5.2c-.4.1-.6.1-.6.1h-.1c-1.2.2-2.3-.5-2.5-1.7c-.2-1.2.5-2.3 1.7-2.5c.2.1.3.1.4.1z\"/><path fill=\"#EDC0A2\" d=\"M258.4 91.4c-1.4 0-2.4-1.1-2.4-2.4s1.1-2.4 2.4-2.4c4.4 0 8-3.6 8-8s-3.6-8-8-8c-1.4 0-2.4-1.1-2.4-2.4s1.1-2.4 2.4-2.4c7.1 0 12.9 5.8 12.9 12.9s-5.8 12.7-12.9 12.7z\"/><path fill=\"#2B3B47\" d=\"M129.6 476.7c-4.1 0-8.3-1.3-11.9-4c-6.7-4.8-9.8-12.8-8-20.7c.2-1.1 1.3-1.7 2.4-1.5c1.1.2 1.7 1.3 1.5 2.4c-1.4 6.3 1 12.7 6.4 16.6c5.4 3.9 12.2 4.2 17.7.9c.9-.6 2.2-.3 2.7.7c.6.9.3 2.2-.7 2.7c-3.1 1.9-6.6 2.9-10.1 2.9z\"/><path fill=\"#76DFFF\" d=\"M139.7 473.8c-3.2 1.9-6.7 2.8-10.2 2.8c-4.1 0-8.3-1.3-11.9-4c-6.7-4.8-9.8-12.8-8-20.7c.2-1.1 1.3-1.7 2.4-1.5c1.1.2 1.7 1.3 1.5 2.4c-.7 3.2-.4 6.4.7 9.3c4.2-.6 7.7-2.2 5.7-6.5c-7.5-16-7.9-.8-24.9-.4c-3.9 0-9.4-.6-17.4-2.3c-13.4-2.8-22.3-10.6-27.4-2.3c.4.4 1.3 1.1 2.7 2.3c.9.8 2 1.7 3.3 2.8c1.3 1.1 2.8 2.2 4.4 3.4c.8.6 1.7 1.3 2.6 1.9c.9.6 1.9 1.3 2.8 2c1.9 1.4 4 2.8 6.1 4.3l.8.6l.1.1l.2.1l.4.3c.6.4 1.1.7 1.7 1.1c1.1.7 2.3 1.5 3.4 2.2c2.3 1.5 4.6 3 7 4.6c2.3 1.5 4.7 2.8 7.1 4.1c2.3 1.3 4.7 2.7 7 4.1c2.3 1.4 4.5 2.6 6.6 3.9c2 1.3 3.9 2.6 5.7 3.8c7 4.7 11.7 7.8 11.7 7.8c3.1-4.3 8.5-11.8 11.5-16c2.9-3.9 5-7.7 4.9-10.7c-.2.4-.3.4-.5.5z\"/><path fill=\"#E5E5E5\" d=\"M49.9 450.5s1.1.9 2.9 2.5c.9.8 2 1.7 3.3 2.8c1.3 1.1 2.8 2.2 4.4 3.4c.8.6 1.7 1.3 2.6 1.9c.9.6 1.9 1.3 2.8 2c1.9 1.4 4 2.8 6.1 4.3l.8.6l.1.1h.1s-.1-.1 0 0l.2.1l.4.3c.6.4 1.1.7 1.7 1.1c1.1.7 2.3 1.5 3.4 2.2c2.3 1.5 4.6 3 7 4.6c2.3 1.5 4.7 2.8 7.1 4.1c2.3 1.3 4.7 2.7 7 4.1c2.3 1.4 4.5 2.6 6.6 3.9c2 1.3 3.9 2.6 5.7 3.8c7 4.7 11.7 7.8 11.7 7.8l-4.7 6.4s-4.5-3.3-11.2-8.1c-1.7-1.2-3.5-2.5-5.4-3.9c-1.9-1.4-4.1-2.7-6.2-4.1c-2.2-1.4-4.4-2.8-6.8-4.3c-2.4-1.5-4.8-3-7.1-4.6c-2.3-1.6-4.5-3.3-6.8-4.9l-3.3-2.4c-.5-.4-1.1-.8-1.6-1.2l-.9-.6l-.8-.6c-2.1-1.6-4.1-3.2-6-4.6c-1-.7-1.9-1.5-2.8-2.1c-.9-.7-1.7-1.4-2.6-2.1c-1.6-1.4-3.2-2.5-4.4-3.7c-1.3-1.2-2.4-2.2-3.3-3c-1.8-1.7-2.8-2.6-2.8-2.6l2.8-3.2z\"/><circle cx=\"99.9\" cy=\"460.8\" r=\"2.2\" fill=\"#00B1FF\"/><circle cx=\"106.7\" cy=\"460.8\" r=\"2.4\" fill=\"#00B1FF\"/><path fill=\"#2B3B47\" d=\"M106.7 461.9c-.5 0-1-.4-1-1c-.1-2.2-.9-6.7-2.8-6.9c-.6-.1-1-.6-.9-1.1c.1-.6.6-1 1.1-.9c4.2.4 4.7 8 4.7 8.9c0 .5-.5 1-1.1 1zm-6.8 0c-.5 0-1-.4-1-1c-.1-2.2-.9-6.7-2.8-6.9c-.6-.1-1-.6-.9-1.1c.1-.6.6-1 1.1-.9c4.2.4 4.7 8 4.7 8.9c0 .5-.5.9-1.1 1z\"/><path fill=\"#00B1FF\" d=\"M110.7 491.6c.4.3.9.6 1.3.9c7 4.7 11.7 7.8 11.7 7.8c3.1-4.3 8.5-11.8 11.5-16c.2-.3.5-.6.7-1c-9.1-2.9-19.3.2-25.2 8.3z\"/><path fill=\"#FFD3B6\" d=\"M253 204.9c-1.3-2.3-2.2-4.8-2.7-7.3c-2.1 2.2-4.2 4.5-6.3 6.7l-14.9 16l-4.9 5.3l-47.1-38.7c.7-1.8 1-3.8.7-5.9l-.6-4.2c-.2-1.4-.4-2.8-.7-4.2c-.2-1.4-.5-2.8-.7-4.2c-.2-1.4-.5-2.8-.8-4.2c-1.1-5.3-6-9.1-11.5-8.8c-.4 0-.8.1-1.2.1c-4-4-10.5-4.2-14.8-.4c-2.2 2-4.5 3.9-6.7 6l-1 .9c-4 3.7-4.4 9.9-.8 14.1c2.7 3.2 5.5 6.4 8.4 9.4c.7.8 1.5 1.5 2.2 2.3c.7.8 1.5 1.5 2.2 2.3c1.5 1.5 3.1 3 4.7 4.4c.6.5 1.3.9 2 1.2l56.4 62.7l.1.1c5.4 5.9 14.6 6.3 20.5.8l1.7-1.6l14.9-15.8l15-15.7l7.8-8.1c-8.8-.1-17.3-4.9-21.9-13.2zm104.9-8.4l-1.7-1l-36.2-19.2l-8.3-4.5l2.6 33.6l18.7 9.7l-14.2 47.8l2.8 35.8c.1.4.1.8.1 1.1c0 2.5-1.3 4.9-3.6 7.1l2.7 6.8c1.6-1.1 3-2.7 3.9-4.7l8.8-20.5c.6-1.5 1-3.3 1-5c0-2.1-.5-4.1-1.4-5.8l29.8-64.9c2.7-5.8.6-12.9-5-16.3zM182.5 362.6c-.9 1.1-1.7 2.3-2.6 3.4l-5.7 7.8l-.2.2l-.1.1v.1l.1-.1l-.4.5l-.7 1l-1.5 2l-2.9 3.9c-2 2.6-3.9 5.2-5.8 7.9c-1.9 2.6-3.8 5.2-5.7 7.9c-1.8 2.6-3.7 5.3-5.5 7.9c-1.8 2.6-3.5 5.2-5.2 7.8c-3.4 5.1-6.7 10.1-9.7 15c-1.5 2.4-3 4.8-4.3 7c-1.4 2.3-2.7 4.5-4 6.7c-1.3 2.1-2.5 4.2-3.6 6.2s-2.2 3.9-3.2 5.7c-1 1.8-1.9 3.5-2.7 5c-.8 1.6-1.6 3-2.2 4.3c-.4.7-.7 1.4-1 2.1c1.1 1.7 2.6 3.3 4.3 4.6c4.6 3.3 10.3 4.1 15.3 2.1c.9-1 1.9-1.9 3-3c2.8-2.8 6-6 9.4-9.5c6.8-7 14.7-15.2 23.2-23.9c8.4-8.8 17.3-18.3 26.3-27.7c2.3-2.4 4.6-4.8 6.9-7.3c2.2-2.4 4.3-4.8 6.5-7.2c.7-.8 1.4-1.5 2-2.2l-30-28.3z\"/><path fill=\"#FFF\" d=\"M88.5 473.2c4.1 2.6 8.7 3.8 13.2 3.8c5.3 0 10.6-1.7 15-5c-1.1-.9-2-1.8-2.9-2.8c-7.1 5-16.8 5.2-24.2 0c.9-4.8.3-9.8-1.6-14.4c-1.5-.2-3-.4-4.8-.7c3.6 6.3 3.8 14.3.1 20.8c.8.5 1.6 1 2.4 1.6c.3.2.6.4 1 .6c.7-1.3 1.3-2.6 1.8-3.9z\"/><path fill=\"#68442A\" d=\"M267 0c-21.5 0-51.3 4.9-52.3 5.7c-1 .8 2 6.9 1 7.8c-2 1.9-3.9 4-5.9 6.4c-1.8 2.2 1 3.5 10.2 3.8c2.4.1-.4 6 2.4 6c1.5 0 28.9-12.1 28.9 31.2c0 3.4 1.6 5.6 4.8 5.8c.4-.7 1.2-1.2 2.1-1.2c7.1 0 12.9 5.8 12.9 12.9c0 1.7-.3 3.3-.9 4.8c6.5 10.7 32.2 19.5 33.2 18.5c7.2-7.7 14.3-27.6 14.3-47.4C317.9 34.6 304.3 0 267 0z\"/><path fill=\"#FFD3B6\" d=\"M425.4 408.6c1.4-4.5 5.1-3.4 8.5-1.1c1-2.8 1.3-6 .6-9.2c-1.4-6.5-6.2-11.3-12.6-12.5c-.7-.1-1.2-.6-1.4-1.2l-75 7.7l-4.7-9.9c-.1 0-.2.1-.2.1l-38.9 14l12.5 26.7l1.2 2.6c3.4 7.3 11.6 11.3 19.7 9.2l89.5-23.3c.1-.8.4-1.9.8-3.1z\"/><path fill=\"#76DFFF\" d=\"M473.5 455.4c-.1-1.2-.4-2.6-.6-4.3c-.2-1.7-.6-3.5-1-5.5c-.2-1-.4-2.1-.6-3.2c-.2-1.1-.5-2.2-.8-3.4c-.5-2.3-1.1-4.8-1.7-7.3l-.2-1v-.1l-.1-.2l-.1-.5c-.2-.6-.3-1.3-.5-1.9c-.3-1.3-.7-2.6-1-3.9l-2.1-8.1c-.8-2.7-1.7-5.2-2.5-7.8c-.8-2.5-1.7-5.2-2.5-7.7s-1.6-5-2.2-7.3c-.6-2.3-1.2-4.6-1.7-6.6c-2.1-8.2-3.5-13.6-3.5-13.6c-5.2 1.1-14.2 3.1-19.2 4.2c-4.8 1-8.8 2.4-10.8 4.7h.4c7.9 1.5 14 7.5 15.7 15.6c1.8 8.1-1.3 16-7.9 20.7c-.3.2-.8.4-1.1.4c-.6 0-1.2-.3-1.6-.8c-.6-.9-.4-2.1.5-2.8c2.7-1.9 4.6-4.5 5.7-7.4c-3.5-2.4-7.2-3.5-8.6 1.1c-5.1 16.9 5.3 5.8 18.2 16.9c2.9 2.6 6.6 6.7 11.5 13.3c8.2 11 9.5 22.8 18.9 20.1c-.2-.6-.4-1.8-.6-3.6z\"/><path fill=\"#E5E5E5\" d=\"M474 459.2s-.2-1.4-.5-3.8c-.1-1.2-.4-2.6-.6-4.3c-.2-1.7-.6-3.5-1-5.5c-.2-1-.4-2.1-.6-3.2c-.2-1.1-.5-2.2-.8-3.4c-.5-2.3-1.1-4.8-1.7-7.3l-.2-1v-.2v.1l-.1-.2l-.1-.5c-.2-.6-.3-1.3-.5-1.9c-.3-1.3-.7-2.6-1-3.9l-2.1-8.1c-.8-2.7-1.7-5.2-2.5-7.8c-.8-2.5-1.7-5.2-2.5-7.7s-1.6-5-2.2-7.3c-.6-2.3-1.2-4.6-1.7-6.6c-2.1-8.2-3.5-13.6-3.5-13.6l7.8-1.7s1.2 5.4 2.9 13.5c.4 2 .9 4.2 1.4 6.6c.5 2.3 1.2 4.7 1.9 7.2s1.3 5 2.1 7.8c.7 2.7 1.6 5.4 2.3 8.2l1.8 8.1c.3 1.3.6 2.7.9 4c.1.7.3 1.3.4 2l.2 1.1l.2 1c.5 2.6 1 5.1 1.4 7.5c.2 1.2.4 2.3.6 3.5c.2 1.1.3 2.2.5 3.3c.3 2.1.7 4 .8 5.7c.2 1.7.4 3.2.5 4.4c.2 2.4.3 3.8.3 3.8l-4.4.2z\"/><circle cx=\"443.7\" cy=\"418.2\" r=\"2.2\" fill=\"#00B1FF\"/><circle cx=\"438.6\" cy=\"413.6\" r=\"2.4\" fill=\"#00B1FF\"/><path fill=\"#2B3B47\" d=\"M436.1 423.1c-.3 0-.6-.1-.8-.3c-2.9-3.1 1.9-9.1 2.4-9.8c.4-.4 1-.5 1.5-.1c.4.4.5 1 .1 1.5c-1.4 1.7-3.8 5.6-2.5 7.1c.4.4.4 1.1-.1 1.5c-.1 0-.3.1-.6.1zm5.1 4.5c-.3 0-.6-.1-.8-.3c-2.9-3.1 1.9-9.1 2.4-9.8c.4-.4 1-.5 1.5-.1s.5 1 .1 1.5c-1.4 1.7-3.8 5.6-2.5 7.1c.4.4.4 1.1-.1 1.5c-.1 0-.4.1-.6.1zm-11.8-9c-.6 0-1.2-.3-1.6-.8c-.6-.9-.4-2.1.5-2.8c5.3-3.8 7.7-10.1 6.3-16.6s-6.2-11.3-12.6-12.5c-1.1-.2-1.8-1.2-1.6-2.3c.2-1.1 1.3-1.8 2.3-1.6c7.9 1.5 14 7.5 15.7 15.6c1.8 8.1-1.3 16-7.9 20.7c-.3.1-.7.3-1.1.3z\"/><path fill=\"#00B1FF\" d=\"m456.3 388.1l-1.2-15.7c-3.8.8-15.9 3.4-22 4.8l-1.2.3c4.8 8.1 14.5 12.7 24.4 10.6z\"/><path fill=\"#FFF\" d=\"M464.7 416c-.1-.4-.2-.7-.3-1.1c-1.4.5-2.7 1-4 1.7c-2.9-10.1-11.7-17-21.8-17.9c.2 1.4.3 2.7.3 4c3.1.4 6.2 1.4 8.9 3.2c4.7 3 7.9 7.6 9.1 13c-3.9 3-6.8 7.1-8.4 11.8c1 1.1 2 2.3 3.1 3.7c1.5-7.1 6.7-13.2 13.9-15.6c-.3-1-.6-1.9-.8-2.8z\"/><path fill=\"#00B1FF\" d=\"m344.4 374l-26.3-67.1c-8.6 8.3-31.7 14.2-58.9 14.2c-22.6 0-36.7-4.1-43.8-10.2l-36.7 39.2c-2.5 2.6-2.4 6.8.3 9.3l36 33.9c2.6 2.5 6.8 2.4 9.3-.3l30.3-32.6c3.3-3.5 9.2-2.4 10.9 2.1l13.8 35.3c1.3 3.3 5 5 8.4 3.8l52.8-19c3.5-1.3 5.3-5.2 3.9-8.6z\"/><path fill=\"#597B91\" d=\"m321.6 298.6l-10.1-129c0-14.5-13.9-26.8-33.4-31.7v7.8c.2.6.3 1.1.3 1.7c0 5.3-8.6 9.6-19.2 9.6s-19.2-4.3-19.2-9.6v-8.8c-14.4 5.3-19.9 17.2-19.9 31l-3.9 49.4l7.9 6.5l4.9-5.3l14.9-16c2.1-2.2 4.2-4.5 6.3-6.7c.5 2.5 1.4 4.9 2.7 7.3c4.6 8.3 13.1 13 22 13.2l-7.8 8.1l-15 15.7l-14.9 15.8l-1.7 1.6c-5.9 5.4-15.1 5.1-20.5-.8l-.1-.1l-1.5-1.7l-3.2 40.9c-.1.7-.2 1.4-.2 2.1c0 11.8 14.8 21.3 49.3 21.3s62.5-9.5 62.5-21.3c-.1-.3-.1-.6-.2-1z\"/>"}},"lastModified":1656182158,"width":512,"height":512});
-Iconify.addCollection({"prefix":"flat-color-icons","icons":{"donate":{"body":"<path fill=\"#E69329\" d=\"m11.7 21.6l5.1 9.9l9.5-3.9l4.4-12.7l-14.8.8z\"/><circle cx=\"15\" cy=\"36\" r=\"7.8\" fill=\"#546E7A\"/><g fill=\"#90A4AE\"><path d=\"M15 27c-5 0-9 4-9 9s4 9 9 9s9-4 9-9s-4-9-9-9zm0 16c-3.9 0-7-3.1-7-7s3.1-7 7-7s7 3.1 7 7s-3.1 7-7 7z\"/><path d=\"M14 33h2v8h-2z\"/></g><g fill=\"#FFB74D\"><path d=\"M12.9 36c1 1.9 3.2 2.7 5.1 1.7l16.5-8.5c1-.5 1.7-1.2 2.2-1.9c1.7-3.2 5.6-10.7 8.2-17.2l-18.2 8.7l-4.8 7.2l-6.8 3.6c-2.6 1.3-3.4 4.2-2.2 6.4z\"/><path d=\"M30.2 3L13.7 9.3c-.7.2-1.5 1-2.2 1.7l-5.6 7.5c-1 1.5-1.2 3.4-.5 5.1c.4 1 1.7 3.4 3.1 6.1C10.1 28 12.4 27 15 27c.4 0 .9 0 1.3.1l-2.1-4.2l4.6-4.1h8s15.5-2.2 18.2-8.7L30.2 3z\"/></g><path fill=\"#FFCDD2\" d=\"M18.2 36c-1.3.6-2.8 0-3.3-1.3c-.6-1.3 0-2.8 1.3-3.3c1.2-.6 3.2 4 2 4.6z\"/>"},"clock":{"body":"<circle cx=\"24\" cy=\"24\" r=\"20\" fill=\"#00ACC1\"/><circle cx=\"24\" cy=\"24\" r=\"16\" fill=\"#eee\"/><path d=\"M23 11h2v13h-2z\"/><path d=\"M31.285 29.654L29.66 31.28l-6.504-6.504l1.626-1.627z\"/><circle cx=\"24\" cy=\"24\" r=\"2\"/><circle cx=\"24\" cy=\"24\" r=\"1\" fill=\"#00ACC1\"/>"}},"lastModified":1672652184,"width":48,"height":48});
-Iconify.addCollection({"prefix":"codicon","icons":{"github-inverted":{"body":"<path fill=\"currentColor\" fill-rule=\"evenodd\" d=\"M7.976 0A7.977 7.977 0 0 0 0 7.976c0 3.522 2.3 6.507 5.431 7.584c.392.049.538-.196.538-.392v-1.37c-2.201.49-2.69-1.076-2.69-1.076c-.343-.93-.881-1.175-.881-1.175c-.734-.489.048-.489.048-.489c.783.049 1.224.832 1.224.832c.734 1.223 1.859.88 2.3.685c.048-.538.293-.88.489-1.076c-1.762-.196-3.621-.881-3.621-3.964c0-.88.293-1.566.832-2.153c-.05-.147-.343-.978.098-2.055c0 0 .685-.196 2.201.832c.636-.196 1.322-.245 2.007-.245s1.37.098 2.006.245c1.517-1.027 2.202-.832 2.202-.832c.44 1.077.146 1.908.097 2.104a3.16 3.16 0 0 1 .832 2.153c0 3.083-1.86 3.719-3.62 3.915c.293.244.538.733.538 1.467v2.202c0 .196.146.44.538.392A7.984 7.984 0 0 0 16 7.976C15.951 3.572 12.38 0 7.976 0z\" clip-rule=\"evenodd\"/>"}},"lastModified":1695361735});
-Iconify.addCollection({"prefix":"ant-design","icons":{"check-square-filled":{"body":"<path fill=\"currentColor\" d=\"M880 112H144c-17.7 0-32 14.3-32 32v736c0 17.7 14.3 32 32 32h736c17.7 0 32-14.3 32-32V144c0-17.7-14.3-32-32-32zM695.5 365.7l-210.6 292a31.8 31.8 0 0 1-51.7 0L308.5 484.9c-3.8-5.3 0-12.7 6.5-12.7h46.9c10.2 0 19.9 4.9 25.9 13.3l71.2 98.8l157.2-218c6-8.3 15.6-13.3 25.9-13.3H689c6.5 0 10.3 7.4 6.5 12.7z\"/>"}},"lastModified":1692940620,"width":1024,"height":1024});
-Iconify.addCollection({"prefix":"emojione-v1","icons":{"white-heavy-check-mark":{"body":"<path fill=\"#37b34a\" d=\"M63.792 56.915a6.876 6.876 0 0 1-6.875 6.876H6.875A6.876 6.876 0 0 1 0 56.915V6.875A6.878 6.878 0 0 1 6.875 0h50.042a6.877 6.877 0 0 1 6.875 6.875v50.04z\"/><path fill=\"#f4f4f4\" d=\"M53.867 14.14a4.656 4.656 0 0 0-6.562.514l-20.04 23.437l-10.781-9a4.248 4.248 0 1 0-5.447 6.519l14.444 12.06a4.223 4.223 0 0 0 3.235.946A4.654 4.654 0 0 0 31.895 47l22.483-26.3a4.659 4.659 0 0 0-.515-6.562\"/>"},"alarm-clock":{"body":"<path fill=\"#c2986b\" d=\"M6.128 61.66c-.215-.293.406-2.372.532-2.521c3.271-3.808 3.391-16.06 3.324-17.806c-.164-4.03 1.972-8.106 3.774-7.225c1.74.85 3.104 9.23 3.439 10.907c.827 4.101.503 9.01-.825 12.898c-1.402 4.117-6.821 8.295-10.244 3.747m51.847 0c.214-.293-.406-2.372-.533-2.521c-3.271-3.808-3.39-16.06-3.323-17.806c.163-4.03-1.973-8.106-3.775-7.225c-1.739.85-3.103 9.23-3.438 10.907c-.827 4.101-.503 9.01.825 12.898c1.401 4.117 6.82 8.295 10.244 3.747\"/><path fill=\"#fbb041\" d=\"M2.421 17.334c2.16 2.161 7.245.573 11.367-3.55c4.12-4.12 5.708-9.203 3.548-11.365C15.181.266 6.522-1.72 2.4 2.4c-4.12 4.12-2.134 12.781.021 14.934m58.999 0c-2.159 2.161-7.244.573-11.366-3.55c-4.12-4.12-5.708-9.203-3.549-11.365C48.661.266 57.319-1.72 61.441 2.4c4.12 4.12 2.134 12.781-.021 14.934\"/><path fill=\"#a7a8ac\" d=\"M60.809 33.07c0 15.979-12.95 28.932-28.93 28.932S2.948 49.05 2.948 33.07c0-15.976 12.95-27.1 28.93-27.1s28.931 11.123 28.931 27.1\"/><path fill=\"#e7e6e6\" d=\"M53.53 33.3c0 11.961-9.694 21.656-21.656 21.656S10.218 45.261 10.218 33.3c0-11.958 9.695-20.283 21.656-20.283S53.53 21.342 53.53 33.3\"/><g fill=\"#25333a\"><path d=\"M34.555 35.13c0 1.756-1.395 3.176-3.117 3.176s-3.117-1.42-3.117-3.176V20.204c0-1.751 1.395-3.171 3.117-3.171c1.723 0 3.117 1.42 3.117 3.171V35.13\"/><path d=\"M31.625 38.31c-1.828 0-3.305-.968-3.305-2.161c0-1.19 1.477-2.157 3.305-2.157h15.557c1.827 0 3.308.967 3.308 2.157c0 1.193-1.48 2.161-3.308 2.161H31.625\"/></g>"},"warning":{"body":"<path fill=\"#ffdd15\" d=\"M63.37 53.52C53.982 36.37 44.59 19.22 35.2 2.07a3.687 3.687 0 0 0-6.522 0C19.289 19.22 9.892 36.37.508 53.52c-1.453 2.649.399 6.083 3.258 6.083h56.35c1.584 0 2.648-.853 3.203-2.01c.698-1.102.885-2.565.055-4.075\"/><path fill=\"#1f2e35\" d=\"m28.917 34.477l-.889-13.262c-.166-2.583-.246-4.439-.246-5.565c0-1.534.4-2.727 1.202-3.588c.805-.856 1.863-1.286 3.175-1.286c1.583 0 2.646.551 3.178 1.646c.537 1.102.809 2.684.809 4.751c0 1.215-.066 2.453-.198 3.708l-1.19 13.649c-.129 1.626-.404 2.872-.827 3.739c-.426.871-1.128 1.301-2.109 1.301c-.992 0-1.69-.419-2.072-1.257c-.393-.841-.668-2.12-.833-3.836m3.072 18.217c-1.125 0-2.106-.362-2.947-1.093c-.841-.728-1.26-1.748-1.26-3.058c0-1.143.4-2.12 1.202-2.921c.805-.806 1.786-1.206 2.951-1.206s2.153.4 2.977 1.206c.815.801 1.234 1.778 1.234 2.921c0 1.29-.419 2.308-1.246 3.044a4.245 4.245 0 0 1-2.911 1.107\"/>"},"backhand-index-pointing-right":{"body":"<path fill=\"#dda37c\" d=\"M29.685 46.14c1.227-2.991 2.176-5.559 2.592-9.403c.829-3.469.727-5.401.553-7.906c-.458-6.568-1.728-8.268-3.866-12.682c-.291-.059-1.377-.145-1.598.033c-.295.238-.569.232-.865.467c-.403.324-.491.965-.237 1.41c-.3.604-.087 1.451.691 1.52c-.025.078-.038.166-.055.244c-.732.25-1.448.563-2.176 1.066c-.587.402-.495 1.281.013 1.689c.461.373.944.623 1.439.811c-.142.17-.274.342-.404.537c-.486.738.017 1.771.945 1.684c.229 2.484.694 5.03 1.069 6.882c-.354 3.302-.271 6.607-.762 9.891c-.521 1.647-1.01 3.283-1.01 4.973v.012c0 .061-.687.108-.837.191c-.753.412-1.02 1.249-.42 1.827c.046.046-.009.063.037.104c-.637.408-.653 1.627.15 1.902c.091.028.141.054.225.078c.008.005 0 .009.004.009a.296.296 0 0 0 .059.017c.232.075.961.059 1.169.112c1.095-1.917 2.452-3.442 3.285-5.468\"/><path fill=\"#efb086\" d=\"M29.859 45.27a32.552 32.552 0 0 1-1.748 3.429l-1.835 3.108c1.194.225 2.396.537 3.583 1.04c.549.225 1.456.241 2.347.038c1.414-.333 1.992-1.598 2.229-2.601c.271-1.17.033-1.914-.948-2.896c-.145-.07-3.628-2.118-3.628-2.118m34.102-28.333c-.07-2.246-1.123-3.52-3.129-3.787c-2.371-.32-8.359-.387-10.635-.115c-2.63.316-5.301.766-7.89 1.201c-2.525.43-5.139.871-7.733 1.191c-2.082.262-4.175.41-6.259.535l2.109 3.867c.745 1.363 1.323 2.928 1.731 4.646l1.282-.686c.217-.117.803-.43 4.76-.629c1.905-.113 4.46-.217 7.082-.32c3.208-.125 6.516-.258 8.858-.416l.882-.059c2.039-.133 4.145-.266 6.104-.736c1.273-.303 2.95-1.118 2.838-4.692M34.627 36.805l-2.254-.388a35.94 35.94 0 0 1-1.994 7.474l.982.149c1.318.204 2.604.313 3.728.313c2.643 0 4.314-.579 4.594-1.586c.383-1.39.283-2.563-.3-3.482c-.779-1.22-2.422-2.081-4.756-2.48m7.023-8.489c-.208-1.418-1.053-1.689-2.65-2.084c-.441-.113-1.885-.201-2.938-.264c-.98-.063-1.896-.119-2.301-.186l-1.389-.234c.479 2.539.616 5.352.358 8.314c-.037.437-.129.856-.179 1.29l2.035-.441c.391-.083 1.057-.171 1.738-.254c.62-.079 1.561-.195 1.811-.278c.508-.167 1.156-.396 1.71-.737c1.514-.923 2.155-2.746 1.805-5.126M27.612 16c-.986.053-6.508-.105-7.84-.105c-1.331 0-1.406-1.057-.47-1.068c.94-.008 5.729-.063 5.729-.063c-1.398-3.799-3.911-3.799-5.667-3.799c-.732 0-1.515.121-2.322.35c-4.951 1.422-8.787 3.67-11.718 6.873C.435 23.514-1.075 29.563.828 36.171c1.991 6.92 5.979 11.409 11.857 13.353c2.56.853 5.359 1.198 8.07 1.535l2.321.287a80.78 80.78 0 0 1 1.948.258l1.147-3.753c3.111-10.145 3.831-18.829 2.275-27.325L27.614 16\"/>"}},"lastModified":1672651819,"width":64,"height":64});
-Iconify.addCollection({"prefix":"twemoji","icons":{"light-bulb":{"body":"<path fill=\"#FFD983\" d=\"M29 11.06c0 6.439-5 7.439-5 13.44c0 3.098-3.123 3.359-5.5 3.359c-2.053 0-6.586-.779-6.586-3.361C11.914 18.5 7 17.5 7 11.06C7 5.029 12.285.14 18.083.14C23.883.14 29 5.029 29 11.06z\"/><path fill=\"#CCD6DD\" d=\"M22.167 32.5c0 .828-2.234 2.5-4.167 2.5c-1.933 0-4.167-1.672-4.167-2.5c0-.828 2.233-.5 4.167-.5c1.933 0 4.167-.328 4.167.5z\"/><path fill=\"#FFCC4D\" d=\"M22.707 10.293a.999.999 0 0 0-1.414 0L18 13.586l-3.293-3.293a.999.999 0 1 0-1.414 1.414L17 15.414V26a1 1 0 1 0 2 0V15.414l3.707-3.707a.999.999 0 0 0 0-1.414z\"/><path fill=\"#99AAB5\" d=\"M24 31a2 2 0 0 1-2 2h-8a2 2 0 0 1-2-2v-6h12v6z\"/><path fill=\"#CCD6DD\" d=\"M11.999 32a1 1 0 0 1-.163-1.986l12-2a.994.994 0 0 1 1.15.822a.999.999 0 0 1-.822 1.15l-12 2a.927.927 0 0 1-.165.014zm0-4a1 1 0 0 1-.163-1.986l12-2a.995.995 0 0 1 1.15.822a.999.999 0 0 1-.822 1.15l-12 2a.927.927 0 0 1-.165.014z\"/>"},"gear":{"body":"<path fill=\"#66757F\" d=\"M34 15h-3.362a12.915 12.915 0 0 0-1.582-3.814l2.379-2.379a2 2 0 0 0 0-2.829l-1.414-1.414a2 2 0 0 0-2.828 0l-2.379 2.379A12.924 12.924 0 0 0 21 5.362V2a2 2 0 0 0-2-2h-2a2 2 0 0 0-2 2v3.362a12.915 12.915 0 0 0-3.814 1.582L8.808 4.565a2 2 0 0 0-2.828 0L4.565 5.979a2.002 2.002 0 0 0-.001 2.829l2.379 2.379A12.918 12.918 0 0 0 5.362 15H2a2 2 0 0 0-2 2v2a2 2 0 0 0 2 2h3.362a12.92 12.92 0 0 0 1.582 3.813l-2.379 2.379c-.78.78-.78 2.048.001 2.829l1.414 1.414c.78.78 2.047.78 2.828 0l2.379-2.379a12.889 12.889 0 0 0 3.814 1.582V34a2 2 0 0 0 2 2h2a2 2 0 0 0 2-2v-3.362a12.92 12.92 0 0 0 3.813-1.582l2.379 2.379a2 2 0 0 0 2.828 0l1.414-1.414a2 2 0 0 0 0-2.829l-2.379-2.379a12.889 12.889 0 0 0 1.582-3.814H34a2 2 0 0 0 2-2v-2A2 2 0 0 0 34 15zM18 26a8 8 0 1 1 0-16a8 8 0 0 1 0 16z\"/>"},"frowning-face":{"body":"<path fill=\"#FFCC4D\" d=\"M36 18c0 9.941-8.059 18-18 18c-9.94 0-18-8.059-18-18C0 8.06 8.06 0 18 0c9.941 0 18 8.06 18 18\"/><ellipse cx=\"11.5\" cy=\"14.5\" fill=\"#664500\" rx=\"2.5\" ry=\"3.5\"/><ellipse cx=\"24.5\" cy=\"14.5\" fill=\"#664500\" rx=\"2.5\" ry=\"3.5\"/><path fill=\"#664500\" d=\"M8.665 27.871a.501.501 0 0 0 .635.029c.039-.029 3.922-2.9 8.7-2.9c4.766 0 8.662 2.871 8.7 2.9a.5.5 0 0 0 .729-.657C27.3 27.029 24.212 22 18 22s-9.301 5.028-9.429 5.243a.499.499 0 0 0 .094.628z\"/>"}},"lastModified":1672668672,"width":36,"height":36});
-Iconify.addCollection({"prefix":"noto","icons":{"play-button":{"body":"<path fill=\"#F77E00\" d=\"M116.46 3.96h-104c-4.42 0-8 3.58-8 8v104c0 4.42 3.58 8 8 8h104c4.42 0 8-3.58 8-8v-104c0-4.42-3.58-8-8-8z\"/><path fill=\"#FF9800\" d=\"M110.16 3.96h-98.2a7.555 7.555 0 0 0-7.5 7.5v97.9c-.01 4.14 3.34 7.49 7.48 7.5h98.12c4.14.01 7.49-3.34 7.5-7.48V11.46c.09-4.05-3.13-7.41-7.18-7.5h-.22z\"/><path fill=\"#FFBD52\" d=\"M40.16 12.86c0-2.3-1.6-3-10.8-2.7c-7.7.3-11.5 1.2-13.8 4s-2.9 8.5-3 15.3c0 4.8 0 9.3 2.5 9.3c3.4 0 3.4-7.9 6.2-12.3c5.4-8.7 18.9-10.6 18.9-13.6z\" opacity=\".75\"/><path fill=\"#FAFAFA\" d=\"M43.7 62.21v-25.7a2.258 2.258 0 0 1 3.4-2l43.5 25.7c1.13.72 1.47 2.22.75 3.35c-.19.3-.45.55-.75.75l-43.5 25.6c-1.08.63-2.46.27-3.09-.81c-.21-.36-.32-.77-.31-1.19v-25.7z\"/>"},"stop-button":{"body":"<path fill=\"#F77E00\" d=\"M116.46 3.96h-104c-4.42 0-8 3.58-8 8v104c0 4.42 3.58 8 8 8h104c4.42 0 8-3.58 8-8v-104c0-4.42-3.58-8-8-8z\"/><path fill=\"#FF9800\" d=\"M110.16 3.96h-98.2a7.555 7.555 0 0 0-7.5 7.5v97.9c-.01 4.14 3.34 7.49 7.48 7.5h98.12c4.14.01 7.49-3.34 7.5-7.48V11.46c.09-4.05-3.13-7.41-7.18-7.5h-.22z\"/><path fill=\"#FFBD52\" d=\"M40.16 12.86c0-2.3-1.6-3-10.8-2.7c-7.7.3-11.5 1.2-13.8 4s-2.9 8.5-3 15.3c0 4.8 0 9.3 2.5 9.3c3.4 0 3.4-7.9 6.2-12.3c5.4-8.7 18.9-10.6 18.9-13.6z\" opacity=\".75\"/><path fill=\"#FAFAFA\" d=\"M89.66 91.96h-50.4c-1.55 0-2.8-1.25-2.8-2.8v-50.4c0-1.55 1.25-2.8 2.8-2.8h50.4a2.728 2.728 0 0 1 2.8 2.66v50.54a2.728 2.728 0 0 1-2.66 2.8h-.14z\"/>"}},"lastModified":1672668145,"width":128,"height":128});
-Iconify.addCollection({"prefix":"eva","icons":{"corner-right-down-fill":{"body":"<g id=\"evaCornerRightDownFill0\"><g id=\"evaCornerRightDownFill1\"><path id=\"evaCornerRightDownFill2\" fill=\"currentColor\" d=\"M18.78 14.38a1 1 0 0 0-1.4-.16L14 16.92V8a3 3 0 0 0-3-3H6a1 1 0 0 0 0 2h5a1 1 0 0 1 1 1v8.92l-3.38-2.7a1 1 0 0 0-1.24 1.56l5 4a1 1 0 0 0 1.24 0l5-4a1 1 0 0 0 .16-1.4Z\"/></g></g>"},"checkmark-circle-2-fill":{"body":"<g id=\"evaCheckmarkCircle2Fill0\"><g id=\"evaCheckmarkCircle2Fill1\"><path id=\"evaCheckmarkCircle2Fill2\" fill=\"currentColor\" d=\"M12 2a10 10 0 1 0 10 10A10 10 0 0 0 12 2Zm4.3 7.61l-4.57 6a1 1 0 0 1-.79.39a1 1 0 0 1-.79-.38l-2.44-3.11a1 1 0 0 1 1.58-1.23l1.63 2.08l3.78-5a1 1 0 1 1 1.6 1.22Z\"/></g></g>"}},"lastModified":1672651966,"width":24,"height":24});
-Iconify.addCollection({"prefix":"bi","icons":{"heart-fill":{"body":"<path fill=\"currentColor\" fill-rule=\"evenodd\" d=\"M8 1.314C12.438-3.248 23.534 4.735 8 15C-7.534 4.736 3.562-3.248 8 1.314z\"/>"},"people-fill":{"body":"<path fill=\"currentColor\" d=\"M7 14s-1 0-1-1s1-4 5-4s5 3 5 4s-1 1-1 1H7Zm4-6a3 3 0 1 0 0-6a3 3 0 0 0 0 6Zm-5.784 6A2.238 2.238 0 0 1 5 13c0-1.355.68-2.75 1.936-3.72A6.325 6.325 0 0 0 5 9c-4 0-5 3-5 4s1 1 1 1h4.216ZM4.5 8a2.5 2.5 0 1 0 0-5a2.5 2.5 0 0 0 0 5Z\"/>"}},"lastModified":1695016311});
-Iconify.addCollection({"prefix":"ic","icons":{"round-person":{"body":"<path fill=\"currentColor\" d=\"M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4s-4 1.79-4 4s1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v1c0 .55.45 1 1 1h14c.55 0 1-.45 1-1v-1c0-2.66-5.33-4-8-4z\"/>"},"round-dashboard":{"body":"<path fill=\"currentColor\" d=\"M4 13h6c.55 0 1-.45 1-1V4c0-.55-.45-1-1-1H4c-.55 0-1 .45-1 1v8c0 .55.45 1 1 1zm0 8h6c.55 0 1-.45 1-1v-4c0-.55-.45-1-1-1H4c-.55 0-1 .45-1 1v4c0 .55.45 1 1 1zm10 0h6c.55 0 1-.45 1-1v-8c0-.55-.45-1-1-1h-6c-.55 0-1 .45-1 1v8c0 .55.45 1 1 1zM13 4v4c0 .55.45 1 1 1h6c.55 0 1-.45 1-1V4c0-.55-.45-1-1-1h-6c-.55 0-1 .45-1 1z\"/>"}},"lastModified":1672653513,"width":24,"height":24});
-Iconify.addCollection({"prefix":"bx","icons":{"bxs-bar-chart-alt-2":{"body":"<path d=\"M6 21H3a1 1 0 0 1-1-1v-8a1 1 0 0 1 1-1h3a1 1 0 0 1 1 1v8a1 1 0 0 1-1 1zm7 0h-3a1 1 0 0 1-1-1V3a1 1 0 0 1 1-1h3a1 1 0 0 1 1 1v17a1 1 0 0 1-1 1zm7 0h-3a1 1 0 0 1-1-1V9a1 1 0 0 1 1-1h3a1 1 0 0 1 1 1v11a1 1 0 0 1-1 1z\" fill=\"currentColor\"/>","hidden":true},"code-curly":{"body":"<path fill=\"currentColor\" d=\"M9 22h1v-2h-.989C8.703 19.994 6 19.827 6 16c0-1.993-.665-3.246-1.502-4C5.335 11.246 6 9.993 6 8c0-3.827 2.703-3.994 3-4h1V2H8.998C7.269 2.004 4 3.264 4 8c0 2.8-1.678 2.99-2.014 3L2 13c.082 0 2 .034 2 3c0 4.736 3.269 5.996 5 6zm13-11c-.082 0-2-.034-2-3c0-4.736-3.269-5.996-5-6h-1v2h.989c.308.006 3.011.173 3.011 4c0 1.993.665 3.246 1.502 4c-.837.754-1.502 2.007-1.502 4c0 3.827-2.703 3.994-3 4h-1v2h1.002C16.731 21.996 20 20.736 20 16c0-2.8 1.678-2.99 2.014-3L22 11z\"/>"},"bxs-heart":{"body":"<path d=\"M20.205 4.791a5.938 5.938 0 0 0-4.209-1.754A5.906 5.906 0 0 0 12 4.595a5.904 5.904 0 0 0-3.996-1.558a5.942 5.942 0 0 0-4.213 1.758c-2.353 2.363-2.352 6.059.002 8.412L12 21.414l8.207-8.207c2.354-2.353 2.355-6.049-.002-8.416z\" fill=\"currentColor\"/>","hidden":true}},"aliases":{"bx-code-curly":{"parent":"code-curly"}},"lastModified":1663655035,"width":24,"height":24});
-Iconify.addCollection({"prefix":"fluent","icons":{"data-bar-horizontal-24-filled":{"body":"<path fill=\"currentColor\" d=\"M21 18.251a2.249 2.249 0 0 0-2.25-2.249H5.25a2.249 2.249 0 1 0 0 4.498h13.5A2.249 2.249 0 0 0 21 18.251Zm-4-6.5a2.249 2.249 0 0 0-2.25-2.249h-9.5a2.25 2.25 0 1 0 0 4.498h9.5A2.249 2.249 0 0 0 17 11.751Zm-4-6.5a2.25 2.25 0 0 0-2.25-2.25l-5.5.001a2.25 2.25 0 0 0 0 4.498h5.5A2.25 2.25 0 0 0 13 5.251Z\"/>"},"key-32-filled":{"body":"<path fill=\"currentColor\" d=\"M15 23v1a1 1 0 0 1-1 1h-2v1a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2v-2.586A2 2 0 0 1 4.586 22l7.738-7.739A8 8 0 1 1 20 20h-2v2a1 1 0 0 1-1 1h-2Zm7-11a2 2 0 1 0 0-4a2 2 0 0 0 0 4Z\"/>","width":32,"height":32},"key-24-filled":{"body":"<path fill=\"currentColor\" d=\"M8.95 8.6a6.554 6.554 0 0 1 6.55-6.55c3.596 0 6.55 2.819 6.55 6.45a6.554 6.554 0 0 1-6.55 6.55a6.243 6.243 0 0 1-1.552-.204A1.25 1.25 0 0 1 12.7 16.05h-1.75v1.75c0 .69-.56 1.25-1.25 1.25H7.95v1.25a1.75 1.75 0 0 1-1.75 1.75H3.7a1.75 1.75 0 0 1-1.75-1.75v-2.172c0-.73.29-1.429.806-1.944L8.99 9.948a.275.275 0 0 0 .07-.244A6.386 6.386 0 0 1 8.95 8.6Zm9.3-1.6a1.25 1.25 0 1 0-2.5 0a1.25 1.25 0 0 0 2.5 0Z\"/>"}},"lastModified":1695621060,"width":24,"height":24});
-Iconify.addCollection({"prefix":"ci","icons":{"settings-filled":{"body":"<path fill=\"currentColor\" d=\"M13.82 22h-3.64a1 1 0 0 1-.977-.786l-.407-1.884a8.002 8.002 0 0 1-1.535-.887l-1.837.585a1 1 0 0 1-1.17-.453L2.43 15.424a1.006 1.006 0 0 1 .193-1.239l1.425-1.3a8.1 8.1 0 0 1 0-1.772L2.623 9.816a1.006 1.006 0 0 1-.193-1.24l1.82-3.153a1 1 0 0 1 1.17-.453l1.837.585c.244-.18.498-.348.76-.5c.253-.142.513-.271.779-.386l.408-1.882A1 1 0 0 1 10.18 2h3.64a1 1 0 0 1 .976.787l.412 1.883a8.192 8.192 0 0 1 1.535.887l1.838-.585a1 1 0 0 1 1.169.453l1.82 3.153c.232.407.152.922-.193 1.239l-1.425 1.3a8.1 8.1 0 0 1 0 1.772l1.425 1.3c.345.318.425.832.193 1.239l-1.82 3.153a1 1 0 0 1-1.17.453l-1.837-.585a7.98 7.98 0 0 1-1.534.886l-.413 1.879a1 1 0 0 1-.976.786ZM11.996 8a4 4 0 1 0 0 8a4 4 0 0 0 0-8Z\"/>","hidden":true},"dot-03-m":{"body":"<path fill=\"currentColor\" d=\"M12 16a4 4 0 1 1 0-8a4 4 0 0 1 0 8Z\"/>","hidden":true}},"lastModified":1678257800,"width":24,"height":24});
-Iconify.addCollection({"prefix":"akar-icons","icons":{"chevron-down":{"body":"<path fill=\"none\" stroke=\"currentColor\" stroke-linecap=\"round\" stroke-linejoin=\"round\" stroke-width=\"2\" d=\"m4 9l8 8l8-8\"/>"}},"lastModified":1693459591,"width":24,"height":24});
-Iconify.addCollection({"prefix":"ls","icons":{"logout":{"body":"<path fill=\"currentColor\" d=\"M0 151v391c0 36 15 68 39 93c24 24 55 37 91 37h196v-81H130c-27 0-48-22-48-49V151c0-27 21-48 48-48h196V21H130c-36 0-67 14-91 38c-24 25-39 56-39 92zm215 118v156c0 18 16 33 34 33h181v123c0 11 6 20 16 25c4 1 8 1 10 1c7 0 13-2 18-7l235-235c11-9 10-27 0-37L474 94c-14-15-44-6-44 18v124H249c-18 0-34 15-34 33z\"/>","height":672}},"lastModified":1656182593,"width":717,"height":717});
-Iconify.addCollection({"prefix":"majesticons","icons":{"clipboard-copy":{"body":"<g fill=\"none\"><path fill-rule=\"evenodd\" clip-rule=\"evenodd\" d=\"M10 4a1 1 0 0 0 0 2h4a1 1 0 0 0 0-2h-4zm0-2a3.001 3.001 0 0 0-2.83 2H6a3 3 0 0 0-3 3v12a3 3 0 0 0 3 3h12a3 3 0 0 0 3-3v-4h-8.586l1.293 1.293a1 1 0 0 1-1.414 1.414l-3-3a1 1 0 0 1 0-1.414l3-3a1 1 0 1 1 1.414 1.414L12.414 13H21V7a3 3 0 0 0-3-3h-1.17A3.001 3.001 0 0 0 14 2h-4z\" fill=\"currentColor\"/></g>","hidden":true}},"lastModified":1672653934,"width":24,"height":24});
-Iconify.addCollection({"prefix":"fa-regular","icons":{"calendar-alt":{"body":"<path fill=\"currentColor\" d=\"M148 288h-40c-6.6 0-12-5.4-12-12v-40c0-6.6 5.4-12 12-12h40c6.6 0 12 5.4 12 12v40c0 6.6-5.4 12-12 12zm108-12v-40c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v40c0 6.6 5.4 12 12 12h40c6.6 0 12-5.4 12-12zm96 0v-40c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v40c0 6.6 5.4 12 12 12h40c6.6 0 12-5.4 12-12zm-96 96v-40c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v40c0 6.6 5.4 12 12 12h40c6.6 0 12-5.4 12-12zm-96 0v-40c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v40c0 6.6 5.4 12 12 12h40c6.6 0 12-5.4 12-12zm192 0v-40c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v40c0 6.6 5.4 12 12 12h40c6.6 0 12-5.4 12-12zm96-260v352c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V112c0-26.5 21.5-48 48-48h48V12c0-6.6 5.4-12 12-12h40c6.6 0 12 5.4 12 12v52h128V12c0-6.6 5.4-12 12-12h40c6.6 0 12 5.4 12 12v52h48c26.5 0 48 21.5 48 48zm-48 346V160H48v298c0 3.3 2.7 6 6 6h340c3.3 0 6-2.7 6-6z\"/>","width":448}},"lastModified":1656181869,"width":512,"height":512});
-Iconify.addCollection({"prefix":"ph","icons":{"books-bold":{"body":"<path fill=\"currentColor\" d=\"m235.57 193.73l-33.19-157.8a20 20 0 0 0-23.76-15.48l-46.81 10.06a19.82 19.82 0 0 0-11 6.65A20 20 0 0 0 104 28H56a20 20 0 0 0-20 20v160a20 20 0 0 0 20 20h48a20 20 0 0 0 20-20V90.25l25.62 121.82A20 20 0 0 0 169.15 228a20.27 20.27 0 0 0 4.23-.45l46.81-10.06a20.1 20.1 0 0 0 15.38-23.76ZM148.19 88.65l39-8.38l2.53 12l-39 8.38Zm7.46 35.5l39-8.38l9.16 43.58l-39 8.38Zm24.06-79.39l2.53 12l-39 8.38l-2.53-12ZM60 88h40v80H60Zm40-36v12H60V52ZM60 204v-12h40v12Zm112.29-.76l-2.53-12l39-8.38l2.53 12Z\"/>"}},"lastModified":1678258124,"width":256,"height":256});
-Iconify.addCollection({"prefix":"fa-solid","icons":{"external-link-alt":{"body":"<path fill=\"currentColor\" d=\"M432 320h-32a16 16 0 0 0-16 16v112H64V128h144a16 16 0 0 0 16-16V80a16 16 0 0 0-16-16H48a48 48 0 0 0-48 48v352a48 48 0 0 0 48 48h352a48 48 0 0 0 48-48V336a16 16 0 0 0-16-16ZM488 0H360c-21.37 0-32.05 25.91-17 41l35.73 35.73L135 320.37a24 24 0 0 0 0 34L157.67 377a24 24 0 0 0 34 0l243.61-243.68L471 169c15 15 41 4.5 41-17V24a24 24 0 0 0-24-24Z\"/>"}},"lastModified":1656181888,"width":512,"height":512});
-Iconify.addCollection({"prefix":"simple-icons","icons":{"wakatime":{"body":"<path fill=\"currentColor\" d=\"M12 0C5.373 0 0 5.373 0 12s5.373 12 12 12s12-5.373 12-12S18.627 0 12 0zm0 2.824a9.176 9.176 0 1 1 0 18.352a9.176 9.176 0 0 1 0-18.352zm5.097 5.058c-.327 0-.61.19-.764.45c-1.025 1.463-2.21 3.162-3.288 4.706l-.387-.636a.897.897 0 0 0-.759-.439a.901.901 0 0 0-.788.492l-.357.581l-1.992-2.943a.897.897 0 0 0-.761-.446c-.514 0-.903.452-.903.96a1 1 0 0 0 .207.61l2.719 3.96c.152.272.44.47.776.47a.91.91 0 0 0 .787-.483c.046-.071.23-.368.314-.504l.324.52c-.035-.047.076.113.087.13c.024.031.054.059.078.085c.019.019.04.036.058.052c.036.033.08.056.115.08c.025.016.052.028.076.04c.029.015.06.024.088.035c.058.025.122.027.18.04c.031.004.064.003.092.005c.29 0 .546-.149.707-.36c1.4-2 2.842-4.055 4.099-5.849A.995.995 0 0 0 18 8.842c0-.508-.389-.96-.903-.96\"/>"}},"lastModified":1695621373,"width":24,"height":24});
-Iconify.addCollection({"prefix":"heroicons-solid","icons":{"light-bulb":{"body":"<path fill=\"currentColor\" d=\"M11 3a1 1 0 1 0-2 0v1a1 1 0 1 0 2 0V3Zm4.657 2.757a1 1 0 0 0-1.414-1.414l-.707.707a1 1 0 0 0 1.414 1.414l.707-.707ZM18 10a1 1 0 0 1-1 1h-1a1 1 0 1 1 0-2h1a1 1 0 0 1 1 1ZM5.05 6.464A1 1 0 1 0 6.464 5.05l-.707-.707a1 1 0 0 0-1.414 1.414l.707.707ZM5 10a1 1 0 0 1-1 1H3a1 1 0 1 1 0-2h1a1 1 0 0 1 1 1Zm3 6v-1h4v1a2 2 0 1 1-4 0Zm4-2c.015-.34.208-.646.477-.859a4 4 0 1 0-4.954 0c.27.213.462.519.476.859h4.002Z\"/>"},"server":{"body":"<path fill=\"currentColor\" fill-rule=\"evenodd\" d=\"M2 5a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v2a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5Zm14 1a1 1 0 1 1-2 0a1 1 0 0 1 2 0ZM2 13a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v2a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2v-2Zm14 1a1 1 0 1 1-2 0a1 1 0 0 1 2 0Z\" clip-rule=\"evenodd\"/>"}},"lastModified":1661525143,"width":20,"height":20});
-Iconify.addCollection({"prefix":"ion","icons":{"rocket":{"body":"<path fill=\"currentColor\" d=\"M328.85 156.79a26.69 26.69 0 1 0 18.88 7.81a26.6 26.6 0 0 0-18.88-7.81Z\"/><path fill=\"currentColor\" d=\"M477.44 50.06a.29.29 0 0 1 0-.09a20.4 20.4 0 0 0-15.13-15.3c-29.8-7.27-76.68.48-128.63 21.28c-52.36 21-101.42 52-134.58 85.22A320.7 320.7 0 0 0 169.55 175c-22.33-1-42 2.18-58.57 9.41c-57.74 25.41-74.23 90.44-78.62 117.14a25 25 0 0 0 27.19 29h.13l64.32-7.02c.08.82.17 1.57.24 2.26a34.36 34.36 0 0 0 9.9 20.72l31.39 31.41a34.27 34.27 0 0 0 20.71 9.91l2.15.23l-7 64.24v.13A25 25 0 0 0 206 480a25.25 25.25 0 0 0 4.15-.34C237 475.34 302 459.05 327.34 401c7.17-16.46 10.34-36.05 9.45-58.34a314.78 314.78 0 0 0 33.95-29.55c33.43-33.26 64.53-81.92 85.31-133.52c20.69-51.36 28.48-98.59 21.39-129.53ZM370.38 224.94a58.77 58.77 0 1 1 0-83.07a58.3 58.3 0 0 1 0 83.07Z\"/><path fill=\"currentColor\" d=\"M161.93 386.44a16 16 0 0 0-11 2.67c-6.39 4.37-12.81 8.69-19.29 12.9c-13.11 8.52-28.79-6.44-21-20l12.15-21a16 16 0 0 0-15.16-24.91A61.25 61.25 0 0 0 72 353.56c-3.66 3.67-14.79 14.81-20.78 57.26A357.94 357.94 0 0 0 48 447.59A16 16 0 0 0 64 464h.4a359.87 359.87 0 0 0 36.8-3.2c42.47-6 53.61-17.14 57.27-20.8a60.49 60.49 0 0 0 17.39-35.74a16 16 0 0 0-13.93-17.82Z\"/>"}},"lastModified":1679641094,"width":512,"height":512});
-Iconify.addCollection({"prefix":"mdi","icons":{"language-c":{"body":"<path fill=\"currentColor\" d=\"m15.45 15.97l.42 2.44c-.26.14-.68.27-1.24.39c-.57.13-1.24.2-2.01.2c-2.21-.04-3.87-.7-4.98-1.96c-1.14-1.27-1.68-2.88-1.68-4.83C6 9.9 6.68 8.13 8 6.89C9.28 5.64 10.92 5 12.9 5c.75 0 1.4.07 1.94.19s.94.25 1.2.4l-.6 2.49l-1.04-.34c-.4-.1-.87-.15-1.4-.15c-1.15-.01-2.11.36-2.86 1.1c-.76.73-1.14 1.85-1.18 3.34c.01 1.36.37 2.42 1.08 3.2c.71.77 1.7 1.17 2.99 1.18l1.33-.12c.43-.08.79-.19 1.09-.32Z\"/>"},"language-cpp":{"body":"<path fill=\"currentColor\" d=\"m10.5 15.97l.41 2.44c-.26.14-.68.27-1.24.39c-.57.13-1.24.2-2.01.2c-2.21-.04-3.87-.7-4.98-1.96C1.56 15.77 1 14.16 1 12.21c.05-2.31.72-4.08 2-5.32C4.32 5.64 5.96 5 7.94 5c.75 0 1.4.07 1.94.19s.94.25 1.2.4l-.58 2.49l-1.06-.34c-.4-.1-.86-.15-1.39-.15c-1.16-.01-2.12.36-2.87 1.1c-.76.73-1.15 1.85-1.18 3.34c0 1.36.37 2.42 1.08 3.2c.71.77 1.71 1.17 2.99 1.18l1.33-.12c.43-.08.79-.19 1.1-.32M11 11h2V9h2v2h2v2h-2v2h-2v-2h-2v-2m7 0h2V9h2v2h2v2h-2v2h-2v-2h-2v-2Z\"/>"},"language-go":{"body":"<path fill=\"currentColor\" d=\"m2.64 10.33l-.02-.06l.22-.27l.12-.08H6.8l.03.08l-.18.26l-.11.06l-3.9.01m-1.61.98L1 11.26l.22-.29l.12-.06h4.9l.05.09l-.08.24l-.1.07H1.03m2.6.99l-.04-.06l.16-.28l.1-.06H6l.07.07l-.02.25l-.08.08H3.63m11.15-2.16l-1.78.47c-.19.04-.2.05-.34-.11c-.16-.18-.27-.29-.5-.4c-.66-.34-1.33-.24-1.94.15c-.72.48-1.11 1.17-1.1 2.05c.01.86.6 1.57 1.45 1.7c.73.09 1.34-.17 1.83-.72l.29-.39h-2.07c-.22 0-.27-.14-.2-.32l.55-1.18c.03-.06.11-.17.27-.17h3.44c.15-.5.41-.96.75-1.41c.78-1.03 1.73-1.57 3-1.81c1.07-.18 2.13-.07 3.07.57c.84.58 1.37 1.36 1.5 2.39c.19 1.45-.24 2.63-1.24 3.65c-.71.72-1.58 1.17-2.57 1.39l-.86.08c-.98-.02-1.87-.3-2.62-.95c-.52-.45-.88-.99-1.06-1.63c-.15.24-.27.47-.44.7c-.77 1.02-1.78 1.65-3.06 1.8c-1.05.16-2.03-.05-2.89-.69c-.79-.6-1.26-1.4-1.36-2.39c-.14-1.17.2-2.22.91-3.14c.76-1 1.77-1.63 3.01-1.86c1-.18 1.97-.06 2.84.52c.57.38.97.9 1.24 1.52c.04.09 0 .15-.12.18m6.11 1.6l-.03-.36c-.19-1.06-1.17-1.66-2.19-1.43c-1.01.22-1.67.85-1.88 1.86c-.19.84.21 1.69.98 2.03c.59.26 1.19.22 1.76-.06c.84-.43 1.31-1.12 1.36-2.04Z\"/>"},"language-haskell":{"body":"<path fill=\"currentColor\" d=\"m2.08 19l4.67-7l-4.67-7h3.5l4.67 7l-4.67 7h-3.5m4.67 0l4.67-7l-4.67-7h3.5l9.34 14h-3.5l-2.92-4.37L10.25 19h-3.5m11.28-4.08l-1.53-2.34h5.42v2.34h-3.89m-2.33-3.5l-1.56-2.34h7.78v2.34H15.7Z\"/>"},"language-html5":{"body":"<path fill=\"currentColor\" d=\"m12 17.56l4.07-1.13l.55-6.1H9.38L9.2 8.3h7.6l.2-1.99H7l.56 6.01h6.89l-.23 2.58l-2.22.6l-2.22-.6l-.14-1.66h-2l.29 3.19L12 17.56M4.07 3h15.86L18.5 19.2L12 21l-6.5-1.8L4.07 3Z\"/>"},"language-java":{"body":"<path fill=\"currentColor\" d=\"M16.5 6.08s-6.84 1.71-3.56 5.48c.97 1.11-.25 2.11-.25 2.11s2.45-1.25 1.31-2.85c-1.06-1.47-1.86-2.2 2.5-4.74m-4.47 1.2C16.08 4.08 14 2 14 2c.84 3.3-2.96 4.3-4.33 6.36c-.94 1.4.46 2.91 2.33 4.64c-.71-1.7-3.22-3.16.03-5.72M9.37 17.47c-3.08.86 1.88 2.63 5.79.96c-.38-.15-.75-.33-1.1-.54c-1.36.31-2.76.37-4.14.18c-1.31-.16-.55-.6-.55-.6m5.32-1.68c-1.75.38-3.56.47-5.34.26c-1.31-.13-.45-.77-.45-.77c-3.4 1.13 1.88 2.4 6.6 1.02c-.29-.11-.57-.3-.81-.51m3.42 3.3s.57.47-.61.83c-2.28.68-9.43.89-11.41.03c-.71-.31.63-.74 1.05-.83c.23-.06.46-.08.69-.08c-.79-.54-5.13 1.1-2.19 1.56c7.97 1.3 14.54-.6 12.47-1.51m-2.74-4.86c.29-.19.6-.35.92-.49c0 0-1.51.26-3.02.4c-1.6.16-3.21.18-4.81.06c-2.35-.31 1.29-1.2 1.29-1.2c-1.1 0-2.18.26-3.16.75c-2.05 1 5.1 1.45 8.78.48m.9 2.42c-.02.04-.04.07-.08.1c5.01-1.31 3.17-4.64.77-3.81c-.13.06-.24.14-.31.25c.14-.05.28-.09.43-.12c1.2-.24 2.92 1.63-.81 3.58m.13 4.61c-3.01.52-6.09.56-9.12.14c0 0 .46.38 2.81.53c3.6.23 9.13-.13 9.26-1.83c.03.01-.23.65-2.95 1.16Z\"/>"},"language-javascript":{"body":"<path fill=\"currentColor\" d=\"M3 3h18v18H3V3m4.73 15.04c.4.85 1.19 1.55 2.54 1.55c1.5 0 2.53-.8 2.53-2.55v-5.78h-1.7V17c0 .86-.35 1.08-.9 1.08c-.58 0-.82-.4-1.09-.87l-1.38.83m5.98-.18c.5.98 1.51 1.73 3.09 1.73c1.6 0 2.8-.83 2.8-2.36c0-1.41-.81-2.04-2.25-2.66l-.42-.18c-.73-.31-1.04-.52-1.04-1.02c0-.41.31-.73.81-.73c.48 0 .8.21 1.09.73l1.31-.87c-.55-.96-1.33-1.33-2.4-1.33c-1.51 0-2.48.96-2.48 2.23c0 1.38.81 2.03 2.03 2.55l.42.18c.78.34 1.24.55 1.24 1.13c0 .48-.45.83-1.15.83c-.83 0-1.31-.43-1.67-1.03l-1.38.8Z\"/>"},"language-kotlin":{"body":"<path fill=\"currentColor\" d=\"M2 2h20L12 12l10 10H2Z\"/>"},"language-lua":{"body":"<path fill=\"currentColor\" d=\"M10.5 5A8.5 8.5 0 0 0 2 13.5a8.5 8.5 0 0 0 8.5 8.5a8.5 8.5 0 0 0 8.5-8.5A8.5 8.5 0 0 0 10.5 5m3 8a2.5 2.5 0 0 1-2.5-2.5A2.5 2.5 0 0 1 13.5 8a2.5 2.5 0 0 1 2.5 2.5a2.5 2.5 0 0 1-2.5 2.5m6-11A2.5 2.5 0 0 0 17 4.5A2.5 2.5 0 0 0 19.5 7A2.5 2.5 0 0 0 22 4.5A2.5 2.5 0 0 0 19.5 2\"/>"},"language-php":{"body":"<path fill=\"currentColor\" d=\"M12 18.08c-6.63 0-12-2.72-12-6.08s5.37-6.08 12-6.08S24 8.64 24 12s-5.37 6.08-12 6.08m-5.19-7.95c.54 0 .91.1 1.09.31c.18.2.22.56.13 1.03c-.1.53-.29.87-.58 1.09c-.28.22-.71.33-1.29.33h-.87l.53-2.76h.99m-3.5 5.55h1.44l.34-1.75h1.23c.54 0 .98-.06 1.33-.17c.35-.12.67-.31.96-.58c.24-.22.43-.46.58-.73c.15-.26.26-.56.31-.88c.16-.78.05-1.39-.33-1.82c-.39-.44-.99-.65-1.82-.65H4.59l-1.28 6.58m7.25-8.33l-1.28 6.58h1.42l.74-3.77h1.14c.36 0 .6.06.71.18c.11.12.13.34.07.66l-.57 2.93h1.45l.59-3.07c.13-.62.03-1.07-.27-1.36c-.3-.27-.85-.4-1.65-.4h-1.27L12 7.35h-1.44M18 10.13c.55 0 .91.1 1.09.31c.18.2.22.56.13 1.03c-.1.53-.29.87-.57 1.09c-.29.22-.72.33-1.3.33h-.85l.5-2.76h1m-3.5 5.55h1.44l.34-1.75h1.22c.55 0 1-.06 1.35-.17c.35-.12.65-.31.95-.58c.24-.22.44-.46.58-.73c.15-.26.26-.56.32-.88c.15-.78.04-1.39-.34-1.82c-.36-.44-.99-.65-1.82-.65h-2.75l-1.29 6.58Z\"/>"},"language-python":{"body":"<path fill=\"currentColor\" d=\"M19.14 7.5A2.86 2.86 0 0 1 22 10.36v3.78A2.86 2.86 0 0 1 19.14 17H12c0 .39.32.96.71.96H17v1.68a2.86 2.86 0 0 1-2.86 2.86H9.86A2.86 2.86 0 0 1 7 19.64v-3.75a2.85 2.85 0 0 1 2.86-2.85h5.25a2.85 2.85 0 0 0 2.85-2.86V7.5h1.18m-4.28 11.79c-.4 0-.72.3-.72.89c0 .59.32.71.72.71a.71.71 0 0 0 .71-.71c0-.59-.32-.89-.71-.89m-10-1.79A2.86 2.86 0 0 1 2 14.64v-3.78A2.86 2.86 0 0 1 4.86 8H12c0-.39-.32-.96-.71-.96H7V5.36A2.86 2.86 0 0 1 9.86 2.5h4.28A2.86 2.86 0 0 1 17 5.36v3.75a2.85 2.85 0 0 1-2.86 2.85H8.89a2.85 2.85 0 0 0-2.85 2.86v2.68H4.86M9.14 5.71c.4 0 .72-.3.72-.89c0-.59-.32-.71-.72-.71c-.39 0-.71.12-.71.71s.32.89.71.89Z\"/>"},"language-r":{"body":"<path fill=\"currentColor\" d=\"M12 4.29c-5.5 0-10 3-10 6.71c0 3.28 3.56 6 8.24 6.58v2.13h3.41v-2.12c.85-.09 1.64-.25 2.39-.48l1.38 2.6h3.86l-2.32-3.91C20.83 14.58 22 12.87 22 11c0-3.71-4.5-6.71-10-6.71m1.53 2.62c4.2 0 7.3 1.4 7.3 4.59c0 1.71-.92 2.91-2.42 3.65c-.09-.05-.17-.1-.22-.15c-.36-.16-.96-.34-.96-.34s2.98-.22 2.98-3.19c0-2.97-3.12-3.02-3.12-3.02h-6.85v7.16c-2.55-.74-4.31-2.31-4.31-4.11c0-2.54 3.4-4.59 7.6-4.59m.15 3.98h2.07s.95-.05.95.94c0 .97-.95.97-.95.97h-2.07v-1.91m-.03 4.41h.92c.18 0 .27.05.43.2c.13.1.27.29.39.46c-.55.07-1.13.1-1.74.1v-.76Z\"/>"},"language-ruby":{"body":"<path fill=\"currentColor\" d=\"M18.8 2.07c2.52.43 3.24 2.16 3.2 3.97V6l-1.14 14.93l-14.78 1.01h.01c-1.23-.05-3.96-.17-4.09-3.99l1.37-2.5l2.77 6.46l2.36-7.62l-.05.01l.02-.02l7.71 2.46l-1.99-7.78l7.35-.46l-5.79-4.74l3.05-1.7v.01M2 17.91v.02v-.02M6.28 6.23c2.96-2.95 6.79-4.69 8.26-3.2c1.46 1.47-.08 5.09-3.04 8.03c-3 2.94-6.77 4.78-8.24 3.3c-1.47-1.49.04-5.19 3.01-8.13h.01Z\"/>"},"language-rust":{"body":"<path fill=\"currentColor\" d=\"m21.9 11.7l-.9-.5V11l.7-.7c.1-.1.1-.3 0-.4l-.1-.1l-.9-.3c0-.1 0-.2-.1-.2l.6-.8c.1-.1.1-.3-.1-.4c0 0-.1 0-.1-.1l-1-.2c0-.1-.1-.1-.1-.2l.4-.9v-.3c-.1-.1-.2-.1-.3-.1h-1s0-.1-.1-.1l.2-1c0-.2-.1-.3-.2-.3h-.1l-1 .2c0-.1-.1-.1-.2-.2v-1c0-.2-.1-.3-.3-.3h-.1l-.9.4h-.1L16 3c0-.2-.2-.3-.3-.2h-.1l-.8.6c-.1 0-.2 0-.2-.1l-.3-.9c-.1-.1-.2-.2-.4-.2c0 0-.1 0-.1.1L13 3h-.2l-.5-.8c-.1-.2-.3-.2-.5-.2l-.1.1l-.5.9H11l-.7-.7c-.1-.1-.3-.1-.4 0l-.1.1l-.3.9c-.1 0-.2 0-.2.1l-.8-.6c-.2-.1-.4-.1-.5.1V3l-.2 1s-.1 0-.2.1l-.9-.4c-.1-.1-.3 0-.4.1v1.1c0 .1-.1.1-.1.2l-1-.2c-.2-.1-.3 0-.3.2v.1l.2 1c-.1 0-.1.1-.2.1h-1c-.2 0-.3.1-.3.3v.1l.4.9v.2L3 8c-.2 0-.3.2-.3.3v.1l.6.8c0 .1 0 .2-.1.2l-.8.4c-.1.1-.2.2-.2.4c0 0 0 .1.1.1l.7.7v.2l-.8.5c-.2.1-.2.3-.2.4l.1.1l.9.6v.2l-.7.7c-.1.1-.1.3 0 .4l.1.1l.9.3c0 .1 0 .2.1.2l-.6.8c-.1.1-.1.3.1.4c0 0 .1 0 .1.1l1 .2c0 .1.1.1.1.2l-.4.9c-.1.1 0 .3.1.4h1.1c.1 0 .1.1.2.1l-.2 1c0 .2.1.3.2.3h.1l1-.2c0 .1.1.1.2.2v1c0 .2.1.3.3.3h.1l.9-.4h.1l.2 1c0 .2.2.3.3.2h.1l.8-.6c.1 0 .2 0 .2.1l.3.9c.1.1.2.2.4.2c0 0 .1 0 .1-.1l.8-.7h.2l.5.8c.1.1.3.2.4.1l.1-.1l.5-.8h.2l.7.7c.1.1.3.1.4 0l.1-.1l.3-.9c.1 0 .2 0 .2-.1l.8.6c.1.1.3.1.4-.1c0 0 0-.1.1-.1l.2-1c.1 0 .1-.1.2-.1l.9.4c.1.1.3 0 .4-.1v-1.1l.2-.2l1 .2c.2 0 .3-.1.3-.2v-.1l-.2-1l.2-.2h1c.2 0 .3-.1.3-.3v-.1l-.4-.9c0-.1.1-.1.1-.2l1-.2c.2 0 .3-.2.2-.3v-.1l-.6-.8l.1-.2l.9-.3c.1-.1.2-.2.2-.4c0 0 0-.1-.1-.1L21 13v-.2l.8-.5c.2-.1.2-.3.1-.6c0 .1 0 .1 0 0m-5.7 7c-.3-.1-.5-.4-.5-.7c.1-.3.4-.5.7-.5c.3.1.5.4.5.7c0 .4-.3.6-.7.5m-.2-1.9c-.3-.1-.6.1-.6.4l-.4 1.4c-.9.4-1.9.6-3 .6s-2.1-.2-3.1-.7l-.3-1.4c-.1-.3-.3-.5-.6-.4l-1.2.3c-.2-.2-.4-.5-.6-.7h6c.1 0 .1 0 .1-.1v-2.1c0-.1 0-.1-.1-.1h-1.7v-1.3h1.9c.2 0 .9 0 1.2 1c.1.3.2 1.3.4 1.6c.1.3.6 1 1.1 1h3.1c-.2.3-.4.5-.7.8l-1.5-.3m-8.3 1.9c-.3.1-.6-.1-.7-.5c-.1-.3.1-.6.5-.7s.6.1.7.5c0 .3-.2.6-.5.7M5.4 9.5c.1.3 0 .7-.3.8c-.3.1-.7 0-.8-.3c-.1-.3 0-.7.3-.8c.4-.1.7 0 .8.3m-.7 1.6l1.3-.5c.3-.1.4-.4.3-.7L6 9.3h1V14H5c-.3-1-.4-1.9-.3-2.9m5.6-.4V9.3h2.5c.1 0 .9.1.9.7c0 .5-.6.7-1.1.7h-2.3m9 1.2v.5h-.8c-.1 0-.1 0-.1.1v.3c0 .8-.5 1-.9 1s-.8-.2-.9-.4c-.2-1.3-.6-1.5-1.2-2c.7-.5 1.5-1.2 1.5-2.1c0-1-.7-1.6-1.1-1.9c-.7-.4-1.4-.5-1.6-.5H6.6c1.1-1.2 2.5-2 4.1-2.3l.9 1c.2.2.5.2.8 0l1-1c2.1.4 3.9 1.7 5 3.6l-.7 1.6c-.1.3 0 .6.3.7l1.3.6v.8m-7.7-8c.2-.2.6-.2.8 0c.2.2.2.6 0 .8c-.3.3-.6.3-.9 0c-.2-.2-.1-.5.1-.8m6.9 5.6c.1-.3.5-.4.8-.3c.3.1.4.5.3.8c-.1.3-.5.4-.8.3c-.3-.1-.4-.5-.3-.8Z\"/>"},"language-swift":{"body":"<path fill=\"currentColor\" d=\"M17.09 19.72c-2.36 1.36-5.59 1.5-8.86.1A13.807 13.807 0 0 1 2 14.5c.67.55 1.46 1 2.3 1.4c3.37 1.57 6.73 1.46 9.1 0c-3.37-2.59-6.24-5.96-8.37-8.71c-.45-.45-.78-1.01-1.12-1.51c8.28 6.05 7.92 7.59 2.41-1.01c4.89 4.94 9.43 7.74 9.43 7.74c.16.09.25.16.36.22c.1-.25.19-.51.26-.78c.79-2.85-.11-6.12-2.08-8.81c4.55 2.75 7.25 7.91 6.12 12.24c-.03.11-.06.22-.05.39c2.24 2.83 1.64 5.78 1.35 5.22c-1.21-2.39-3.48-1.65-4.62-1.17Z\"/>"},"language-typescript":{"body":"<path fill=\"currentColor\" d=\"M3 3h18v18H3V3m10.71 14.86c.5.98 1.51 1.73 3.09 1.73c1.6 0 2.8-.83 2.8-2.36c0-1.41-.81-2.04-2.25-2.66l-.42-.18c-.73-.31-1.04-.52-1.04-1.02c0-.41.31-.73.81-.73c.48 0 .8.21 1.09.73l1.31-.87c-.55-.96-1.33-1.33-2.4-1.33c-1.51 0-2.48.96-2.48 2.23c0 1.38.81 2.03 2.03 2.55l.42.18c.78.34 1.24.55 1.24 1.13c0 .48-.45.83-1.15.83c-.83 0-1.31-.43-1.67-1.03l-1.38.8M13 11.25H8v1.5h1.5V20h1.75v-7.25H13v-1.5Z\"/>"},"language-markdown":{"body":"<path fill=\"currentColor\" d=\"M20.56 18H3.44C2.65 18 2 17.37 2 16.59V7.41C2 6.63 2.65 6 3.44 6h17.12c.79 0 1.44.63 1.44 1.41v9.18c0 .78-.65 1.41-1.44 1.41M6.81 15.19v-3.66l1.92 2.35l1.92-2.35v3.66h1.93V8.81h-1.93l-1.92 2.35l-1.92-2.35H4.89v6.38h1.92M19.69 12h-1.92V8.81h-1.92V12h-1.93l2.89 3.28L19.69 12Z\"/>"},"vuejs":{"body":"<path fill=\"currentColor\" d=\"M2 3h3.5L12 15l6.5-12H22L12 21L2 3m4.5 0h3L12 7.58L14.5 3h3L12 13.08L6.5 3Z\"/>"},"react":{"body":"<path fill=\"currentColor\" d=\"M12 10.11c1.03 0 1.87.84 1.87 1.89c0 1-.84 1.85-1.87 1.85c-1.03 0-1.87-.85-1.87-1.85c0-1.05.84-1.89 1.87-1.89M7.37 20c.63.38 2.01-.2 3.6-1.7c-.52-.59-1.03-1.23-1.51-1.9a22.7 22.7 0 0 1-2.4-.36c-.51 2.14-.32 3.61.31 3.96m.71-5.74l-.29-.51c-.11.29-.22.58-.29.86c.27.06.57.11.88.16l-.3-.51m6.54-.76l.81-1.5l-.81-1.5c-.3-.53-.62-1-.91-1.47C13.17 9 12.6 9 12 9c-.6 0-1.17 0-1.71.03c-.29.47-.61.94-.91 1.47L8.57 12l.81 1.5c.3.53.62 1 .91 1.47c.54.03 1.11.03 1.71.03c.6 0 1.17 0 1.71-.03c.29-.47.61-.94.91-1.47M12 6.78c-.19.22-.39.45-.59.72h1.18c-.2-.27-.4-.5-.59-.72m0 10.44c.19-.22.39-.45.59-.72h-1.18c.2.27.4.5.59.72M16.62 4c-.62-.38-2 .2-3.59 1.7c.52.59 1.03 1.23 1.51 1.9c.82.08 1.63.2 2.4.36c.51-2.14.32-3.61-.32-3.96m-.7 5.74l.29.51c.11-.29.22-.58.29-.86c-.27-.06-.57-.11-.88-.16l.3.51m1.45-7.05c1.47.84 1.63 3.05 1.01 5.63c2.54.75 4.37 1.99 4.37 3.68c0 1.69-1.83 2.93-4.37 3.68c.62 2.58.46 4.79-1.01 5.63c-1.46.84-3.45-.12-5.37-1.95c-1.92 1.83-3.91 2.79-5.38 1.95c-1.46-.84-1.62-3.05-1-5.63c-2.54-.75-4.37-1.99-4.37-3.68c0-1.69 1.83-2.93 4.37-3.68c-.62-2.58-.46-4.79 1-5.63c1.47-.84 3.46.12 5.38 1.95c1.92-1.83 3.91-2.79 5.37-1.95M17.08 12c.34.75.64 1.5.89 2.26c2.1-.63 3.28-1.53 3.28-2.26c0-.73-1.18-1.63-3.28-2.26c-.25.76-.55 1.51-.89 2.26M6.92 12c-.34-.75-.64-1.5-.89-2.26c-2.1.63-3.28 1.53-3.28 2.26c0 .73 1.18 1.63 3.28 2.26c.25-.76.55-1.51.89-2.26m9 2.26l-.3.51c.31-.05.61-.1.88-.16c-.07-.28-.18-.57-.29-.86l-.29.51m-2.89 4.04c1.59 1.5 2.97 2.08 3.59 1.7c.64-.35.83-1.82.32-3.96c-.77.16-1.58.28-2.4.36c-.48.67-.99 1.31-1.51 1.9M8.08 9.74l.3-.51c-.31.05-.61.1-.88.16c.07.28.18.57.29.86l.29-.51m2.89-4.04C9.38 4.2 8 3.62 7.37 4c-.63.35-.82 1.82-.31 3.96a22.7 22.7 0 0 1 2.4-.36c.48-.67.99-1.31 1.51-1.9Z\"/>"},"code-json":{"body":"<path fill=\"currentColor\" d=\"M5 3h2v2H5v5a2 2 0 0 1-2 2a2 2 0 0 1 2 2v5h2v2H5c-1.07-.27-2-.9-2-2v-4a2 2 0 0 0-2-2H0v-2h1a2 2 0 0 0 2-2V5a2 2 0 0 1 2-2m14 0a2 2 0 0 1 2 2v4a2 2 0 0 0 2 2h1v2h-1a2 2 0 0 0-2 2v4a2 2 0 0 1-2 2h-2v-2h2v-5a2 2 0 0 1 2-2a2 2 0 0 1-2-2V5h-2V3h2m-7 12a1 1 0 0 1 1 1a1 1 0 0 1-1 1a1 1 0 0 1-1-1a1 1 0 0 1 1-1m-4 0a1 1 0 0 1 1 1a1 1 0 0 1-1 1a1 1 0 0 1-1-1a1 1 0 0 1 1-1m8 0a1 1 0 0 1 1 1a1 1 0 0 1-1 1a1 1 0 0 1-1-1a1 1 0 0 1 1-1Z\"/>"},"bash":{"body":"<path fill=\"currentColor\" d=\"M5 9h2.31l.32-3h2l-.32 3h2l.32-3h2l-.32 3H15v2h-1.9l-.2 2H15v2h-2.31l-.32 3h-2l.32-3h-2l-.32 3h-2l.32-3H5v-2h1.9l.2-2H5V9m4.1 2l-.2 2h2l.2-2M19 6h-2v8h2m0 2h-2v2h2Z\"/>"},"filter":{"body":"<path fill=\"currentColor\" d=\"M14 12v7.88c.04.3-.06.62-.29.83a.996.996 0 0 1-1.41 0l-2.01-2.01a.989.989 0 0 1-.29-.83V12h-.03L4.21 4.62a1 1 0 0 1 .17-1.4c.19-.14.4-.22.62-.22h14c.22 0 .43.08.62.22a1 1 0 0 1 .17 1.4L14.03 12H14Z\"/>"}},"lastModified":1689058119,"width":24,"height":24});
-Iconify.addCollection({"prefix":"jam","icons":{"crown-f":{"body":"<path fill=\"currentColor\" d=\"M2.049 1.802L5.854 5.15L9.244.976a1 1 0 0 1 1.565.017l3.235 4.156l3.928-3.396a1 1 0 0 1 1.643.9L18.115 13H1.922L.399 2.7a1 1 0 0 1 1.65-.898zM2 14h16v1a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1v-1z\"/>","top":-4}},"lastModified":1656182518,"left":-2,"top":-2,"width":24,"height":24});
-Iconify.addCollection({"prefix":"octicon","icons":{"project-16":{"body":"<path fill=\"currentColor\" d=\"M1.75 0h12.5C15.216 0 16 .784 16 1.75v12.5A1.75 1.75 0 0 1 14.25 16H1.75A1.75 1.75 0 0 1 0 14.25V1.75C0 .784.784 0 1.75 0ZM1.5 1.75v12.5c0 .138.112.25.25.25h12.5a.25.25 0 0 0 .25-.25V1.75a.25.25 0 0 0-.25-.25H1.75a.25.25 0 0 0-.25.25ZM11.75 3a.75.75 0 0 1 .75.75v7.5a.75.75 0 0 1-1.5 0v-7.5a.75.75 0 0 1 .75-.75Zm-8.25.75a.75.75 0 0 1 1.5 0v5.5a.75.75 0 0 1-1.5 0ZM8 3a.75.75 0 0 1 .75.75v3.5a.75.75 0 0 1-1.5 0v-3.5A.75.75 0 0 1 8 3Z\"/>"},"share-16":{"body":"<path fill=\"currentColor\" d=\"M3.75 6.5a.25.25 0 0 0-.25.25v6.5c0 .138.112.25.25.25h8.5a.25.25 0 0 0 .25-.25v-6.5a.25.25 0 0 0-.25-.25h-1a.75.75 0 0 1 0-1.5h1c.966 0 1.75.784 1.75 1.75v6.5A1.75 1.75 0 0 1 12.25 15h-8.5A1.75 1.75 0 0 1 2 13.25v-6.5C2 5.784 2.784 5 3.75 5h1a.75.75 0 0 1 0 1.5ZM7.823.177a.25.25 0 0 1 .354 0l2.896 2.896a.25.25 0 0 1-.177.427H8.75v5.75a.75.75 0 0 1-1.5 0V3.5H5.104a.25.25 0 0 1-.177-.427Z\"/>"}},"lastModified":1694583006});
