+++ static/src/app.css
@@ -1,6 +1,5 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+@import 'virtual:uno.css';
+@import 'npm:@fontsource/source-sans-3';
 
 /* Base definition */
 body {
