+++ views/signup.tpl.html
@@ -7,10 +7,6 @@
     // Constants
     const defaultAvatarUrl = 'assets/images/unknown.svg'
     const avatarUrlTemplate = {{ avatarUrlTemplate }}
-
-    function guessTimezone() {
-        return Intl.DateTimeFormat().resolvedOptions().timeZone
-    }
 </script>
 <script type="module" src="assets/js/components/signup.js"></script>
